% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{analytePointPattern}
\alias{analytePointPattern}
\title{analytePointPattern Class Constructor}
\usage{
analytePointPattern(spp = NA, x = NA, y = NA, intensity = NA, win = NA, mzVals)
}
\arguments{
\item{spp}{a spatial point pattern of type 'spatstat::ppp'.}

\item{x, y, intensity, win}{optional, either these arguments or 'spp' have to be supplied.}

\item{mzVals}{a numeric vector holding the m/z values that represent the analyte under study.}
}
\value{
An S3 object of type 'ppp' and 'analytePointPattern'.
}
\description{
This is a class to extend 'spatstat::ppp' class only enforcing specific point-marks. It also
adds an 'mzVals' slot to declare which analytes are constructing a given 'analytePointPattern' object.
It represents the spatial point pattern of an analyte (or set of
analytes) with marks being a data.frame of one column c("intensity").
}
