% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_Functions.R
\name{angle}
\alias{angle}
\title{Calculate the angle between two vectors (degrees)}
\usage{
angle(x, y)
}
\arguments{
\item{x}{the first vector}

\item{y}{the second vector}
}
\value{
angle in degrees
}
\description{
Calculate the angle between two vectors (degrees)
}
\keyword{internal}
