% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{app2im}
\alias{app2im}
\title{Convert `anlaytePointPattern` to `im`}
\usage{
app2im(obj, rescale = FALSE, zero.rm = FALSE)
}
\arguments{
\item{obj}{S3 object of type `anlaytePointPattern` (and `spp`).}

\item{rescale}{logical, whether to scale the intensities of the output `im` object
to [0,1] interval.}

\item{zero.rm}{for internal use only.}
}
\value{
an object of type `im` (see `spatstat` documentation).
}
\description{
A method to convert `anlaytePointPattern` objects to `im` objects.
}
\keyword{internal}
