% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_Functions.R
\name{atom.pairs.num}
\alias{atom.pairs.num}
\title{Helper function for manipulating human input}
\usage{
atom.pairs.num(atom_pairs)
}
\arguments{
\item{atom_pairs}{A character of atoms (even number of atoms)}
}
\value{
list of numeric atom pairs
}
\description{
Breaks a string of atoms, separated by spaces, into pairs
}
\keyword{internal}
