% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{atom.vectors}
\alias{atom.vectors}
\title{Retrieve movement vectors for an atom}
\usage{
atom.vectors(file)
}
\arguments{
\item{file}{a .feather file}
}
\value{
data frame of movement vectors for a single atom
}
\description{
Retrieve movement vectors for an atom
}
\keyword{internal}
