% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_angles_bonds.R
\name{atoms.distance.df}
\alias{atoms.distance.df}
\title{Compute atom distances of a set of molecules}
\usage{
atoms.distance.df(atom_pairs)
}
\arguments{
\item{atom_pairs}{a character}
}
\value{
(num of molecules)x(num of pairs) data frame of the distances in Å
}
\description{
This function acts on a read feather file
Input is a character of atoms, splited to pairs
atoms separated by a space (e.g. '1 2 3 4') are accounted
for as the pairs 1-2 and 3-4
}
