% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz_steRimol.R
\name{b1.scanner}
\alias{b1.scanner}
\title{scans for B1 vectors}
\usage{
b1.scanner(i, substi)
}
\arguments{
\item{i}{degree}

\item{substi}{truncated xyz}
}
\value{
list of B1 vectors
}
\description{
scans for B1 vectors
}
\keyword{internal}
