% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLipidSearch.R
\name{batchLipidSearch}
\alias{batchLipidSearch}
\title{Batch Analyte detection in an MSI dataset}
\usage{
batchLipidSearch(
  spData,
  fwhmObj,
  sldb,
  adduct = c("M-H", "M+H", "M+Na", "M+K"),
  numCores = 1L,
  wMethod = "Gaussian",
  verifiedMasses = NA,
  confirmedOnly = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{spData}{an S3 object of type `sparseIntensityMatrix` holding the sparse MSI data.}

\item{fwhmObj}{an S3 object of type `fwhm` with the estimated fwhm data.}

\item{sldb}{the SwissLipid database loaded as a data.frame.}

\item{adduct}{a character vector specifying the ionisation status of interest, c("M-H", "M+H", "M+Na", "M+K").}

\item{numCores}{an integer, the number of cores to be used for the search, passed to `parallel::mclapply`. Not
supported on Windows machines.}

\item{wMethod}{wighting method; c("Gaussian", "sum", "max", "mean").}

\item{verifiedMasses}{an optional numeric vector of m/z values that are (externally) verified to
be real molecular entities with a certain confidence, ex. 'mz' column of a METASPACE
annotation result.}

\item{confirmedOnly}{if `TRUE`, returns detections only if confirmed by `verifiedMasses`.}

\item{metaData}{optional named list with additional identifiers for the analyte
under study, ex. list(lipidID = "..", sumformular = "..", ..). This will be passed to the `metaData`
slot of the resulting `analytePointPattern` object.}
}
\value{
An analyte point patter of type `ppp` and `analytePointPattern` containing all lipid hits identified in the
MSI dataset `spData` for the specified `adduct` formations taking into account the `fwhm` information.
}
\description{
This applies `moleculaR::searchAnalyte` on a given MSI dataset against the entire SwissLipids database
taking into account the different ionization statuses (i.e. adducts).
}
