% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{bend.vib.df}
\alias{bend.vib.df}
\title{Get bending vibrations frequencies for a set of molecules}
\usage{
bend.vib.df(atom_pairs = NA)
}
\arguments{
\item{atom_pairs}{a character of atom pairs}
}
\value{
A data frame with bending frequencies
}
\description{
Get bending vibrations frequencies for a set of molecules
}
