% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{bend.vib.set}
\alias{bend.vib.set}
\title{Get bending vibrations frequencies}
\usage{
bend.vib.set(atom_pairs)
}
\arguments{
\item{atom_pairs}{a character of atom pairs}
}
\value{
A data frame with bending frequencies
}
\description{
Get bending vibrations frequencies
}
\keyword{internal}
