% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGaussBW.R
\name{calcGaussBW}
\alias{calcGaussBW}
\title{Calculate Gaussian Bandwidth}
\usage{
calcGaussBW(
  spp,
  win,
  weighted = TRUE,
  bw = seq(1, 12, 0.5),
  csr = NULL,
  plot = TRUE
)
}
\arguments{
\item{spp:}{The spatial point pattern.}

\item{win:}{The window object of type `spatstat::owin`.}

\item{weighted:}{Whether the intensities are inlcude into the computation. Switch off for Collective Projections.}

\item{bw:}{A vector, The gaussian band width pool used for Kernel density estimation.}

\item{csr:}{Pre-computed wighted csr model corresponding to the given spp. This
speeds up the computation}

\item{plot:}{Whether to plot.}
}
\value{
A list ..
}
\description{
This function calculates the Gaussian bandwidth to be used for analyte probability maps by
finding the curve infliction point for a curve that represents hotspot area as a function
of bandwidth.
}
