% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{center.of.mass}
\alias{center.of.mass}
\title{Compute molecule's center of mass}
\usage{
center.of.mass(coor_atoms)
}
\arguments{
\item{coor_atoms}{3 atoms character (e.g. '1 2 3')}
}
\value{
center of mass
}
\description{
Compute molecule's center of mass
}
\keyword{internal}
