% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{center.of.mass.substructure}
\alias{center.of.mass.substructure}
\title{Compute molecule's substructure center of mass}
\usage{
center.of.mass.substructure(coor_atoms, sub_atoms)
}
\arguments{
\item{coor_atoms}{3 atoms character (e.g. '1 2 3')}

\item{sub_atoms}{atoms character (e.g. '4 5 6 7 8 9')}
}
\value{
center of mass of a substructure
}
\description{
Compute molecule's substructure center of mass
}
\keyword{internal}
