% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{check.bond.exist}
\alias{check.bond.exist}
\title{verifies that a bond exists}
\usage{
check.bond.exist(atom1, atom2)
}
\arguments{
\item{atom1}{atom 1}

\item{atom2}{atom 2}
}
\value{
logical
}
\description{
verifies that a bond exists
}
\keyword{internal}
