% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_charges.R
\name{cm5.df}
\alias{cm5.df}
\title{Pull CM5 charges for specific atoms}
\usage{
cm5.df(atom_indices, difference_indices = NA)
}
\arguments{
\item{atom_indices}{a character of atoms}

\item{difference_indices}{a character of atom pairs}
}
\value{
A data frame with CM5 charges and differences
}
\description{
Choose atoms for which you wish to have CM5 charges presented.
Charge differences are also available.
}
