% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_charges.R
\name{cm5.info}
\alias{cm5.info}
\title{Pull CM5 charges for specific atoms}
\usage{
cm5.info(atom_index)
}
\arguments{
\item{atom_index}{a character of atoms}
}
\value{
A data frame with CM5 charges
}
\description{
Choose atoms for which you wish to have CM5 charges presented
}
\keyword{internal}
