% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_Functions.R
\name{coor.trans.file}
\alias{coor.trans.file}
\title{Transform coordinate system for a file of choice}
\usage{
coor.trans.file(coor_atoms, molecule)
}
\arguments{
\item{coor_atoms}{3 atoms as character, separated by spaces (e.g. '1 2 3'),
atom 1 - origin, atom 2 - y direction, atom 3 - xy plane}

\item{molecule}{An xyz file to work on}
}
\value{
a new xyz file with the suffix _tc added
}
\description{
Works directly
on xyz files, creates a new file with the suffix _tc added
}
