% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{create_adjacency_list}
\alias{create_adjacency_list}
\title{Create adjacency matrix}
\usage{
create_adjacency_list(bond_pairs)
}
\arguments{
\item{bond_pairs}{a two column data frame of bonded atoms}
}
\value{
distance matrix for all atoms in a molecule
}
\description{
Create adjacency matrix
}
\keyword{internal}
