% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{createHotspotWin}
\alias{createHotspotWin}
\title{Create a Hotspot Window for Simulation}
\usage{
createHotspotWin(shape = "centralCircle", radius = 25, backgroundLength = 100)
}
\description{
Create a Hotspot Window for Simulation
}
