% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{cross_product}
\alias{cross_product}
\title{compute the cross product of two vetors}
\usage{
cross_product(v1, v2)
}
\arguments{
\item{v1}{vector 1}

\item{v2}{vector 2}
}
\value{
cross product vector
}
\description{
compute the cross product of two vetors
}
\keyword{internal}
