% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logistic_models.R
\name{ct.plot}
\alias{ct.plot}
\title{Create a nice classification table (confusion matrix)}
\usage{
ct.plot(class.table)
}
\arguments{
\item{class.table}{a mod.info result}
}
\value{
classification table figure
}
\description{
Takes a mod.info object (x <- mod.info(blah) and then 
ct.plot(x))
}
