% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cube_steRimol.R
\name{cube.steRimol}
\alias{cube.steRimol}
\alias{cube.sterimol}
\title{Compute cube based sterimol parameters}
\usage{
cube.steRimol(
  cubefile,
  coordinates,
  only.sub = T,
  drop = NULL,
  plot = T,
  degrees = 90,
  isovalue = 0.036
)
}
\arguments{
\item{cubefile}{a .cube file}

\item{coordinates}{primary axis, a two atom character}

\item{only.sub}{if TRUE (default) will account only for atoms directly bonded}

\item{drop}{numeric value of an atom, from which onward atoms will be dropped from calculation}

\item{plot}{if TRUE (default) will generate a 3D plot of molecule, with the
density surface and sterimol arrows.}

\item{degrees}{define the rotational scan, defaults to 90 (same as 360)}

\item{isovalue}{what minimal density isovalue should be considered, defaults to 0.036}
}
\value{
sterimol values for a single molecule
}
\description{
This function acts on Gaussian cube file
NOTE - the function is meant to work in a single molecule
data folder and does not allow the choice of a cube file to work with.
This is by design and is due to the workflow approach of moleculaR.
}
