% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{dip.gaussian.df}
\alias{dip.gaussian.df}
\title{Pulls and manipulates dipole moment vector.}
\usage{
dip.gaussian.df(coor_atoms = "")
}
\arguments{
\item{coor_atoms}{A character of atom indices. The last two will define 
the y axis and the xy plane, respecteviley. All indices before those will
define the origin. In cases where a single atom defines the origin, the input
should be 3 indices, in any other case, the centroid of these indices will}
}
\value{
dataframe with npa based DM
}
\description{
Pulls and manipulates dipole moment vector.
}
