% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{dip.gaussian.multi}
\alias{dip.gaussian.multi}
\title{Pulls and manipulates dipole moment vector.
Allows for use of several corrdinate systems.}
\usage{
dip.gaussian.multi(coor_atoms_vector = c(""))
}
\arguments{
\item{coor_atoms_vector}{vector of sets of 3 or 4 atoms character (e.g. c('1 2 3', '4 5 6'))}
}
\value{
data frame with multiple or single DMs
}
\description{
Pulls and manipulates dipole moment vector.
Allows for use of several corrdinate systems.
}
