% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{dip.gaussian}
\alias{dip.gaussian}
\title{Use Gaussain's dipole moment, and perform manipulations to it.}
\usage{
dip.gaussian(coor_atoms = "")
}
\arguments{
\item{coor_atoms}{a character of atom indices. The last two will define 
the y axis and the xy plane, respecteviley. All indices before those will
define the origin. In cases where a single atom defines the origin, the input
should be 3 indices, in any other case, the centroid of these indices will 
define the origin.}
}
\value{
dipole moment
}
\description{
Use Gaussain's dipole moment, and perform manipulations to it.
}
\keyword{internal}
