% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{.bw.scott2}
\alias{.bw.scott2}
\title{Calculate Gaussian bandwideth - `scott`}
\usage{
.bw.scott2(X, isotropic = FALSE, d = NULL)
}
\arguments{
\item{X:}{A point pattern (object of class \code{ppp}).}

\item{isotropic:}{Logical value indicating whether to compute a single bandwidth
for an isotropic Gaussian kernel (isotropic=TRUE) or separate bandwidths for each
coordinate axis (isotropic=FALSE, the default).}
}
\value{
A numeric, estimated Gaussian bandwidth.
geometric mean
}
\description{
This function Calculate Gaussian bandwideth based on Scott's rule of thumb.
Note that this function is already available in recent \code{spatstat} versions (see \code{?.bw.scott}).
This is used internally in `moleculaR::probMap`.
}
