% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{.bw.spAutoCorr}
\alias{.bw.spAutoCorr}
\title{Calculate Gaussian bandwidth based on spatial autocorrelation - `spAutoCor`}
\usage{
.bw.spAutoCorr(sppMoi, bw = seq(1, 10, 1), plot = FALSE, verbose = FALSE)
}
\arguments{
\item{sppMoi:}{The spatial point pattern, object of type `ppp`.}

\item{bw:}{bandwidth steps at which to compute density and the corresponding Moran's I statistic.}

\item{plot:}{whether to plot the result.}

\item{verbose:}{whether to output progress.}
}
\value{
A numeric, the estimated Gaussian bandwidth.
}
\description{
This function computes the Gaussian bandwidth based on sensitivity analysis of spatial autocorrlation.
This is used internally in `moleculaR::probMap`.
}
\keyword{internal}
