% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{.erosion}
\alias{.erosion}
\title{Apply erosion to an spp}
\usage{
.erosion(spp)
}
\arguments{
\item{spp:}{A spatial point pattern (object of class \code{spp}).}
}
\value{
Another `spp` object with its window eroded on the edges with a fixed r = 0.5
length units (pixels).
}
\description{
This function is equivalent to `spatstat.geom::erosion` but applied to the `spp`
object and retains its attributes
}
