% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFwhm.R
\name{.getFwhm}
\alias{.getFwhm}
\title{Full width half Maximum of peaks}
\usage{
.getFwhm(spectrum, peaks)
}
\arguments{
\item{spectrum:}{a spectrum of \code{MALDIquant::MassSpectrum} type.}

\item{peaks:}{a centroided spectrum of \code{MALDIquant::MassPeaks} type.}
}
\value{
Returns a named vector of fwhm values with the same length as the number of peaks within the spectrum.
}
\description{
A method to compute the fwhm for the peaks of a given spectrum.
}
\references{
Adopted from \url{https://gist.github.com/sgibb/3914291}.
}
\keyword{internal}
