% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{.getMaxCurve}
\alias{.getMaxCurve}
\title{Maximum curvature}
\usage{
.getMaxCurve(x, y, plot = FALSE)
}
\arguments{
\item{x:}{a numeric vector.}

\item{plot:}{a logical to plot the result for validation purposes. Defaults to FALSE.}
}
\value{
Returns the maximum curvature threshold (elbow point) of `x`.
}
\description{
Finds the maximum curvature (elbow point) of vector of values representing a curve by finding the maximum distance from the
curve to the line drawn between the peak and tail of that curve. **Deprecated**.
}
\references{
\url{https://en.wikipedia.org/wiki/Unimodal_thresholding}

\url{https://en.wikipedia.org/wiki/Distance_from_a_point_to_a_line}
}
\author{
Denis Abu Sammour, \email{d.abu-sammour@hs-mannheim.de}
}
\keyword{internal}
