% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{.inflictPoint}
\alias{.inflictPoint}
\title{Inflection point of a curve}
\usage{
.inflictPoint(
  x,
  y,
  df = 7,
  adjRange = TRUE,
  xQuery = seq(range(x)[1], range(x)[2], 0.1)
)
}
\arguments{
\item{x:}{x values.}

\item{y:}{y values.}

\item{df:}{degrees of freedom for the smoothing spline.}

\item{adjRange:}{whether to adjust the range of the second derivative to match
the range of y so it could be displayed along the original data.}

\item{xQuery:}{x values, at which to evaluate the fitted spline.
A list ..}
}
\description{
This function calculates the infliction point of a curve based on the 2nd derivative.
This is used internally in `moleculaR::.calcGaussBW`.
}
\keyword{internal}
