% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeIntensity.R
\name{.isNorm}
\alias{.isNorm}
\title{IS Intensity Normalization}
\usage{
.isNorm(x, mz, tol)
}
\arguments{
\item{x:}{a list of `MassPeaks` or `MassSpectrum` objects.}

\item{mz:}{a numeric, the m/z value of the IS.}

\item{tol:}{a numeric, the absolute tolerance for m/z (in Da).}
}
\value{
IS intensity-normalized list of `MassPeaks` or `MassSpectrum` objects.
}
\description{
This normalizes spectral intensities to a provided M/z value i.e. an intenanl
standard (IS).
}
\keyword{internal}
