% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{.kneedle}
\alias{.kneedle}
\title{Find the knee/elbow point in a vector using the Kneedle algorithm.}
\usage{
.kneedle(x, y, sign = 1)
}
\arguments{
\item{x}{numeric vector, x-values.}

\item{y}{numeric vector, y-values.}

\item{sign}{+1 for increasing values and  -1 for decreasing values.}
}
\value{
The index of the knee/elbow.
}
\description{
This function uses the Kneedle algorithm to find the index of the knee point
in the provided x,y-vectors.
}
