% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeIntensity.R
\name{.rmsNorm}
\alias{.rmsNorm}
\title{RMS Intensity Normalization}
\usage{
.rmsNorm(x)
}
\arguments{
\item{x:}{a list of `MassPeaks` or `MassSpectrum` objects.}
}
\value{
RMS intensity-normalized list of `MassPeaks` or `MassSpectrum` objects.

 @keywords internal
}
\description{
This normalizes spectral intensities to their RMS.
}
