% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreateExData.R
\name{.shrinkData}
\alias{.shrinkData}
\title{Shrink data}
\usage{
.shrinkData(x, mzKeep, fwhmObj)
}
\arguments{
\item{x:}{Dataset, a list of \code{MALDIquant::MassPeaks} objects.}

\item{mzKeep:}{numeric vector, m/z values to retain.}

\item{fwhmObj:}{the 'fwhm' S3 object. See \code{?moleculaR::estimateFwhm}.}
}
\value{
Filtered list of \code{MALDIquant::MassPeaks} objects.
}
\description{
This function is for internal use only. It tries to shrink MSI data (in terms of size) for including it within
a package.
}
\keyword{internal}
