% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{.spAutoCorGaussBW}
\alias{.spAutoCorGaussBW}
\title{Calculate Gaussian bandwidth based on spatial autocorrelation - `spAutoCor`}
\usage{
.spAutoCorGaussBW(
  spp,
  win,
  dists = seq(1, 5, 0.5),
  plot = FALSE,
  numSim = 99,
  approximate = FALSE,
  approxRadius = NULL,
  verbose = TRUE
)
}
\arguments{
\item{spp:}{The spatial point pattern.}

\item{win:}{The window object of type `spatstat::owin`.}

\item{dists:}{neighborhood distances at which to compute the Moran's I statisitc..}

\item{plot:}{whether to plot the result.}

\item{numSim:}{the number of Monte Carlo simulations to be done, see ?spdep::moran.mc.}

\item{approximate:}{when `TRUE`only points within a circle centered at the intensity-wighted
centroid (center of gravity) and radius `approxRadius` are considered. This
is used to speed up the caluclation.}

\item{approxRadius:}{the radius in units (usually pixels) for the approximation circlular region.By default,
this equates to the radius that produces a circular area which is half of the area of}

\item{verbose:}{whether to output progress.}
}
\value{
A numeric, the estimated Gaussian bandwidth.
}
\description{
This function computes the Gaussian bandwidth based on sensitivity analysis of spatial autocorrlation.
This is used internally in `moleculaR::probMap`.
}
\keyword{internal}
