% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeIntensity.R
\name{.ticNorm}
\alias{.ticNorm}
\title{TIC Intensity Normalization}
\usage{
.ticNorm(x)
}
\arguments{
\item{x:}{a list of `MassPeaks` or `MassSpectrum` objects.}
}
\value{
RMS intensity-normalized list of `MassPeaks` or `MassSpectrum` objects.
}
\description{
This normalizes spectral intensities to their TIC.
}
\keyword{internal}
