% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadExData.R
\name{downloadExData}
\alias{downloadExData}
\title{Download example MSI data}
\usage{
downloadExData(path = tempdir())
}
\arguments{
\item{path:}{a character string, path to the directory where the downloaded
file is saved. Defaults to \code{tempdir()}. The dataset is save as
\code{moleculaR-example-data.imzML}.}
}
\value{
Returns the path to the downloaded example data file.
}
\description{
This downloads an example dataset for package illustration and development.
}
