% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping-functions.R
\name{dsc}
\alias{dsc}
\title{Dice Similarity Coefficient of two window objects}
\usage{
dsc(win0, win1, winBackground = NULL, plot = FALSE)
}
\arguments{
\item{win0:}{a reference image object of type `owin`.}

\item{win1:}{a second image object of type `owin`.}

\item{winBackground:}{an optional bounding window of type `owin` which acts as
the background window for both `win0` and `win1`. This is only used for
visualization.}

\item{plot:}{whether to plot the result.}
}
\value{
A numeric, the calculated Dice similarity coefficient.
}
\description{
This function Calculates the Dice similarity coefficient of two
window objects of type `owin` (see `?spatstat.geom::owin`).
}
\keyword{internal}
