% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbowPointSimple.R
\name{elbowPointSimple}
\alias{elbowPointSimple}
\title{Calculate analyte probablity map}
\usage{
elbowPointSimple(
  x,
  y,
  df = 7,
  adjRange = TRUE,
  xQuery = seq(range(x)[1], range(x)[2], 0.1)
)
}
\arguments{
\item{x:}{x values.}

\item{y:}{y values.}

\item{df:}{degrees of freedom for the smoothing spline.}

\item{adjRange:}{whether to adjust the range of the second derivative to match
the range of y so it could be displayed along the original data.}

\item{xQuery:}{x values, at which to evaluate the fitted spline.
A list ..}
}
\description{
This function creates an analyte probability map from a given spatial point pattern.
}
\keyword{internal}
