% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFwhm.R
\name{estimateFwhm}
\alias{estimateFwhm}
\title{Estimation of fwhm(m/z)}
\usage{
estimateFwhm(
  s,
  sampling = 1000L,
  plot = FALSE,
  savePlot = NULL,
  returnValues = FALSE
)
}
\arguments{
\item{s:}{a single spectrum or a list of single spectra of 'MALDIquant::MassSpectrum' type. In the latter
case, a single spectrum is randomly chosen to base the computations on.}

\item{sampling:}{an integer specifying how many detected peaks to consider for fwhm estimation, the lower the faster.}

\item{plot:}{whether to plot the result}

\item{savePlot:}{either \code{NULL} or file path to save a plot as svg.}
}
\value{
Returns an S3 object 'fwhm' containing a linear interplotor function 'fwhmInterpolator' in addition to
m/z values of the peaks and their corresponding fwhm values.
}
\description{
This creates a linear interpolator function that approximats fwhm as a function
of m/z for the given single spectrum or a list of mass spectra.
}
