% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractoR.R
\name{extRact.CM5}
\alias{extRact.CM5}
\title{Pull CM5 charges from Gaussian log files}
\usage{
extRact.CM5()
}
\value{
A single column data frame with CM5 charges
}
\description{
This function acts on a read Gaussian log file.
No input is needed.
}
\details{
No Parameters
}
\keyword{internal}
