% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_Functions.R
\name{extract.connectivity}
\alias{extract.connectivity}
\title{Create a bonding data frame}
\usage{
extract.connectivity(
  xyz_file = list.files(pattern = ".xyz"),
  threshold_distance = 2.02
)
}
\arguments{
\item{xyz_file}{An xyz file, automatic for cases in which there's only one.
If used directly, be sure to input or else a non stable behavior is expected.}

\item{threshold_distance}{The upper limit of distance counted as bonded.
Default includes H bonding at 2.02 Å.}
}
\value{
A data frame of bonds
}
\description{
Works directly on xyz files
}
