% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extRactor.R
\name{extRact.Dipole}
\alias{extRact.Dipole}
\title{Pull the dipole moment vector from Gaussian log files}
\usage{
extRact.Dipole()
}
\value{
A 4 column data frame with dipole moment components and magnitude
}
\description{
This function acts on a read Gaussian log file.
No input is needed.
}
\details{
No Parameters
}
\keyword{internal}
