% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extRactor.R
\name{extRact.Hirsh}
\alias{extRact.Hirsh}
\title{Pull Hirshfeld charges from Gaussian log files}
\usage{
extRact.Hirsh()
}
\value{
A single column data frame with Hirshfeld charges
}
\description{
This function acts on a read Gaussian log file.
No input is needed.
}
\details{
No Parameters
}
\keyword{internal}
