% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extRactor.R
\name{extRact.NBO}
\alias{extRact.NBO}
\title{Pull NBO charges from Gaussian log files}
\usage{
extRact.NBO()
}
\value{
A single column data frame with nbo charges
}
\description{
This function acts on a read Gaussian log file.
No input is needed.
}
\details{
No Parameters
}
\keyword{internal}
