% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractoR.R
\name{extRact.polarizability}
\alias{extRact.polarizability}
\title{Pull isotropic and anisotropic polarizability from Gaussian log files}
\usage{
extRact.polarizability()
}
\value{
A 2 column data frame with polarizabilities
}
\description{
This function acts on a read Gaussian log file.
No input is needed.
}
\details{
No Parameters
}
\keyword{internal}
