% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractoR.R
\name{extRact.spectrum}
\alias{extRact.spectrum}
\title{Pull IR spectrum from Gaussian log files}
\usage{
extRact.spectrum()
}
\value{
A 2 column data frame with IR vibrational frequencies and intensities
}
\description{
This function acts on a read Gaussian log file.
No input is needed.
}
\details{
No Parameters
}
\keyword{internal}
