% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractoR.R
\name{extRact.vectors}
\alias{extRact.vectors}
\title{Pull atom movement vectors from Gaussian log files}
\usage{
extRact.vectors()
}
\value{
A 3 column * number of atoms column data frame with movements vectors
}
\description{
This function acts on a read Gaussian log file.
No input is needed.
}
\details{
No Parameters
}
\keyword{internal}
