% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extRactor.R
\name{extRact.xyz}
\alias{extRact.xyz}
\title{Pull xyz of the optimized structure from Gaussian log files}
\usage{
extRact.xyz()
}
\value{
A 4 column data frame with atom symbols and coordinates
}
\description{
This function acts on a read Gaussian log file.
No input is needed.
}
\details{
No Parameters
}
\keyword{internal}
