% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extRactor.R
\name{extRactoR.auto}
\alias{extRactoR.auto}
\title{Extract and compress all needed information from Gaussian log files - no GUI}
\usage{
extRactoR.auto()
}
\value{
A .feather file for each log file
}
\description{
This function acts on a set of Gaussian log files.
No input is needed.
}
\details{
No Parameters
}
