% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extRactor.R
\name{extRactoR}
\alias{extRactoR}
\title{Extract and compress all needed information from Gaussian log files.}
\usage{
extRactoR()
}
\value{
A .feather file for each log file
}
\description{
This function acts on a set of Gaussian log files.
No input is needed.
}
\details{
No Parameters
}
