% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPeaks.R
\name{filterPeaks}
\alias{filterPeaks}
\title{Filter Peaks}
\usage{
filterPeaks(x, minFreq = 0.01)
}
\arguments{
\item{x:}{Dataset, a list of \code{MALDIquant::MassPeaks} objects.}

\item{minFreq:}{numeric falling in [0,1], peaks occuring in \code{<= minFreq * length(x)} will be omitted.}
}
\value{
Filtered list of `MassPeaks` objects.
}
\description{
This function filters out peaks which occur in less than 'minFreq' percent of all pixels.
This is similar to 'MALDIquant::filterPeaks' but is more memory friendly as it relies on
sparse data representation.
}
