% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{find.bend.freq}
\alias{find.bend.freq}
\title{Find a single bending frequency for a pair of atoms}
\usage{
find.bend.freq(atom_pair, threshold = 1350)
}
\arguments{
\item{atom_pair}{a character of two atoms (e.g. '1 2')}

\item{threshold}{minimum frequency value (above fingerprint region)}
}
\value{
frequency
}
\description{
Find a single bending frequency for a pair of atoms
}
\keyword{internal}
