% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{find.center.atom}
\alias{find.center.atom}
\title{Find the atom connecting two atoms, if there is one}
\usage{
find.center.atom(atom1, atom2)
}
\arguments{
\item{atom1}{atom 1 (numeric)}

\item{atom2}{atom 2 (numeric)}
}
\value{
center atom or FALSE
}
\description{
Find the atom connecting two atoms, if there is one
}
\keyword{internal}
