% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{find_circular_paths}
\alias{find_circular_paths}
\title{find_circular_paths - find 6 membered circular paths that include the 
primary ring_vib atom. Used in the process of ring vibrations calculation}
\usage{
find_circular_paths(bonds, node)
}
\arguments{
\item{bonds}{bonds dataframe}

\item{node}{primary ring_vib atom}
}
\value{
6-membered circular paths that include the primary atom
}
\description{
find_circular_paths - find 6 membered circular paths that include the 
primary ring_vib atom. Used in the process of ring vibrations calculation
}
\keyword{internal}
