% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz_steRimol.R
\name{find_paths_with_nodes}
\alias{find_paths_with_nodes}
\title{when only.sub is TRUE.}
\usage{
find_paths_with_nodes(bonds, node1, node2)
}
\arguments{
\item{bonds}{secondary steRimol atom}

\item{node1}{primary steRimol atom}
}
\value{
all simple paths that include the two atoms
}
\description{
when only.sub is TRUE.
}
\keyword{internal}
