% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{focusPlot}
\alias{focusPlot}
\title{Plot a spectrum focusing on a specific mass-range}
\usage{
focusPlot(x, m, fwhm, sigmalim = 6, col = NULL)
}
\arguments{
\item{x}{an object of type `MassPeaks` or `MassSpectrum`. Either a single
spectrum or a mean spectrum (see `?MALDIquant::mergeMassPeaks`).}

\item{m}{a numeric (vector), the m/z value(S) at which to do the plotting.}

\item{fwhm}{a numeric (vector), the fwhm value(s) at `m`, normally generated by `getFwhm`.
Must have the same length as `m`.}

\item{sigmalim}{an integer, how many sigmas (of the calculated Gaussian)
around `m` should be considered in the plot. Note that the superimposed
Gaussian always spans `m ± sigma`.}

\item{col}{a character (vector), colors used for the m/z values in `m`.}
}
\value{
a plot.
}
\description{
This function plots a spectrum (`MassPeaks` or `MassSpectrum`) and provides
a focus on a mass range around a given m/z of interest taking into
consideration the fwhm at that mass.
}
