% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldChangeNorm.R
\name{foldChangeNorm}
\alias{foldChangeNorm}
\title{Normalize Intensity}
\usage{
foldChangeNorm(x, meanMethod = "mean")
}
\arguments{
\item{x:}{a list of \code{MassPeaks} objects.}

\item{meanMethod:}{see \code{?MALDIquant::mergeMassPeaks} for more info.}
}
\value{
Intensity-normalized list of \code{MassPeaks} objects.
}
\description{
This normalizes spectral intensities such that for every spectrum a median
fold-change of peak intensities is calculated with reference to the mean spectrum
which is then taken as a normalization factor for that particular spectrum.
}
\references{
Veselkov, Kirill, et al. "BASIS: High-performance bioinformatics platform for processing of large-scale mass
spectrometry imaging data in chemically augmented histology." Scientific reports 8.1 (2018): 1-11.
}
