% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{fwhmData}
\alias{fwhmData}
\title{fwhmData Class Constructor}
\usage{
fwhmData(fwhmInterpolator, peaks = NA_real_, fwhmVals = NA_real_)
}
\arguments{
\item{fwhmInterpolator}{A function that is used to find fwhm value for any given m/z.}

\item{peaks}{The m/z values of the peaks of a single spectrum.}

\item{fwhmVals}{The corresponding fwhm values in Da.}
}
\value{
An S3 object 'fwhmData'.
}
\description{
This constructs a 'fwhmData' object which stores information about the calculated full width half maximum.
}
