% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{getFwhm}
\alias{getFwhm}
\alias{getFwhm.fwhm}
\title{Get fwhm for a given m/z value}
\usage{
getFwhm(obj, mzVals)

\method{getFwhm}{fwhm}(obj, mzVals)
}
\arguments{
\item{obj}{S3 object of type `fwhm`.}

\item{mzVals}{m/z value (Da) at which to estimate the fwhm.}
}
\value{
returns a numeric, estimated fwhm at the given m/z.
}
\description{
This is a method for S3 class `fwhm` to interpolate fwhm values.
}
