% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_charges.R
\name{hirsh.df}
\alias{hirsh.df}
\title{Pull Hirshfeld charges for specific atoms}
\usage{
hirsh.df(atom_indices, difference_indices = NA)
}
\arguments{
\item{atom_indices}{a character of atoms}

\item{difference_indices}{a character of atom pairs}
}
\value{
A data frame with Hirshfeld charges and differences
}
\description{
Choose atoms for which you wish to have Hirshfeld charges presented.
Charge differences are also available.
}
