% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_charges.R
\name{hirsh.info}
\alias{hirsh.info}
\title{Pull Hirshfeld charges for specific atoms}
\usage{
hirsh.info(atom_index)
}
\arguments{
\item{atom_index}{a character of atoms}
}
\value{
A data frame with Hirshfeld charges
}
\description{
Choose atoms for which you wish to have Hirshfeld charges presented
}
\keyword{internal}
