% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping-functions.R
\name{im2spp}
\alias{im2spp}
\title{Convert `im` to `anlaytePointPattern`}
\usage{
im2spp(obj, win, rescale = FALSE, zero.rm = TRUE, mzVal = "mz")
}
\arguments{
\item{obj}{S3 object of type `im`.}

\item{win}{S3 object of type `owin` representing the window or the resulting
`anlaytePointPattern` object.}

\item{rescale}{logical, whether to scale the intensities of the output `im` object
to [0,1] interval.}

\item{zero.rm}{for internal use only.}

\item{mzVal}{optional, the m/z value of the analyte represnted in `obj`.}
}
\value{
an object of type `anlaytePointPattern`.
}
\description{
A method to convert `im` objects to `anlaytePointPattern` objects.
}
\keyword{internal}
