% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initLipidSearch.R
\name{initLipidSearch}
\alias{initLipidSearch}
\title{Initialize lipids search}
\usage{
initLipidSearch(swissdb)
}
\arguments{
\item{swissdb:}{The swiss lipids db as dataframe.}
}
\value{
a list of two entreis; a list containing segmented lipid
species into individual entries and an additional empty list
with the same structure and naming to store the search results.
}
\description{
This function initializes some lists for lipid search.
}
