% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logistic_models.R
\name{k.fold.log.ordinal}
\alias{k.fold.log.ordinal}
\title{Perform a k-fold cv for an ordinal logistic model}
\usage{
k.fold.log.ordinal(
  formula,
  data,
  folds = NULL,
  outcome.column = which(colnames(data) == "class"),
  stratify = F,
  sample.vector = floor(round(summary(data$class)/min(summary(data$class))))
)
}
\arguments{
\item{formula}{model formula}

\item{data}{a data frame with a class column}

\item{folds}{number of folds (k)}

\item{outcome.column}{column with classes}

\item{stratify}{logical, should stratify in breaking to folds}

\item{sample.vector}{what is the ratio in sampling
between classes (for unbalanced sets) e.g. c(2,2,1)}
}
\value{
model stats
}
\description{
Perform a k-fold cv for an ordinal logistic model
}
\keyword{internal}
