% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{kneePoint}
\alias{kneePoint}
\title{Find Knee (or elbow) point of a curve}
\usage{
kneePoint(
  x,
  y,
  df = 7,
  xQuery = seq(range(x)[1], range(x)[2], 0.1),
  plot = FALSE,
  sign = +1,
  ...
)
}
\arguments{
\item{sign}{+1 for increasing values (knee) and  -1 for decreasing values (elbow).}

\item{x:}{x values representing the bandwidth values}

\item{y:}{y values representing the Moran's I statistic.}

\item{df:}{degrees of freedom for the smoothing spline.}

\item{plot:}{whether to plot the result.}

\item{xQuery:}{x values to be used for smoothing the original curve via
a fitted spline.}

\item{...:}{arguments passed to `plot`.}
}
\value{
Returns a numeric, the calculated knee point representing the optimum bandwidth.
}
\description{
This function calculates the knee/elbow point of a curve based on the kneedle
algorithm (satopaa et al, 2011). This is used internally in `moleculaR::.calcGaussBW`.
This is a simplified implementation.
}
\references{
Satopaa, Ville, et al. "Finding a" kneedle" in a haystack: Detecting knee points
in system behavior." 2011 31st international conference on distributed computing
systems workshops. IEEE, 2011. (doi: 10.1109/ICDCSW.2011.20)
}
\keyword{internal}
