% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{lipidSearchList}
\alias{lipidSearchList}
\title{lipidSearchList Class Constructor}
\usage{
lipidSearchList(lipidList, hitsList, allClasses)
}
\arguments{
\item{lipidList}{a data.frame which lists entries of lipids.}

\item{hitsList}{a data.frame holding the lipids that were detected.}

\item{allClasses}{a character vector summarizing all classes taking part in the lipid search.}
}
\value{
An S3 object of type 'lipidSearchList' .
}
\description{
A class to re-arrange lipid listings in the SwissLipid database to facilitate computational
serach. It splits the full database into lipid class-oriented lists.
}
