% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualSpatstatImport.R
\docType{data}
\name{manualSpatstatImport}
\alias{manualSpatstatImport}
\alias{installedPckgs}
\title{import spatstat sub-packages}
\format{
An object of class \code{character} of length 164.
}
\usage{
installedPckgs
}
\description{
This is done because spatstat is in constant development and on
several occasions some functions jumped from one sub-package to
another (there were also newly created sub-packages). To avoid attaching the entire
namespace of spatstat and its sub-packages (by listing it under "Depends"),
the below loader script is called when the package is loaded
to make it possible to work with tha package without attaching it
via "library()". Note that this adds the requirement of listing
all spatstat functions needed for normal functioning of moleculaR in the
function below.
}
\details{
This is a temporary solution.
}
\keyword{datasets}
