% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logistic_models.R
\name{mod.info.log.ordinal}
\alias{mod.info.log.ordinal}
\title{Model summary - ordinal model}
\usage{
mod.info.log.ordinal(model, data)
}
\arguments{
\item{model}{model object - MASS::polr object}

\item{data}{data frame with class column}
}
\value{
model stats
}
\description{
Model summary - ordinal model
}
