% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logistic_models.R
\name{mod.info.logistic}
\alias{mod.info.logistic}
\title{Model summary - logistic model}
\usage{
mod.info.logistic(model, data)
}
\arguments{
\item{model}{model object - nnet::multinom object}

\item{data}{data frame with class column}
}
\value{
model stats
}
\description{
Model summary - logistic model
}
