% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_models.R
\name{model.cv}
\alias{model.cv}
\title{Cross validate (k-fold) a single model}
\usage{
model.cv(formula, data, out.col, folds, iterations)
}
\arguments{
\item{formula}{model formula (not designed for direct use)}

\item{data}{data frame with outcome column}

\item{out.col}{column number with output}

\item{folds}{define K}

\item{iterations}{Number of iterations}
}
\value{
averaged MAE and Q2
}
\description{
An iterative version of model.single.cv
}
