% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_models.R
\name{model.plot}
\alias{model.plot}
\title{ggplot2 plot of a linear model for QSAR}
\usage{
model.plot(model = models[1, 1], data)
}
\arguments{
\item{model}{model formula (not designed for direct use)}

\item{data}{data frame with outcome column}
}
\value{
a chemistry relevant plot
}
\description{
ggplot2 plot of a linear model for QSAR
}
