% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_GUI.R
\name{model.report.GUI}
\alias{model.report.GUI}
\title{Screen for models, cross validate them and plot. Designed for interactive work.}
\usage{
model.report.GUI()
}
\value{
csv file with model list, and a plot
 use and documentation
}
\description{
Screen for models, cross validate them and plot. Designed for interactive work.
}
