% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logistic_models.R
\name{model.report.logistic}
\alias{model.report.logistic}
\title{Generate a model report - logistic regression}
\usage{
model.report.logistic(dataset, min = 1, max = (min + 1), model.number = 1)
}
\arguments{
\item{dataset}{data frame with class column}

\item{min}{minimum number of variables for model sub-setting}

\item{max}{maximum number of variables for model sub-setting}

\item{model.number}{which model (out of the list) should be summarized}
}
\value{
A full report
}
\description{
A full, detailed report, includes plots
}
