% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parallel_linear_models.R
\name{model.report.parallel}
\alias{model.report.parallel}
\title{Generate models, cross validate, and create visualizations with parallel processing}
\usage{
model.report.parallel(
  dataset,
  min = 2,
  max = NULL,
  leave.out = "",
  predict = FALSE,
  what.model = NULL,
  cor.threshold = 0.7,
  save.pred = TRUE,
  plot_title = "Linear Regression Model Analysis"
)
}
\arguments{
\item{dataset}{A dataframe with outcome column}

\item{min}{Minimum # of features (default = 2)}

\item{max}{Max # of features (defaults = # of observations / 5)}

\item{leave.out}{Name of observations to leave out (e.g. 'p_Br')}

\item{predict}{If leave.out is not empty, should a prediction of it be computed?}

\item{what.model}{If not in an interactive session, what model should be used?}

\item{cor.threshold}{Pearson correlation threshold for variable selection (default = 0.7)}

\item{save.pred}{Logical indicating whether to save out-of-sample predictions to a CSV file (default: TRUE)}

\item{plot_title}{Title for the generated plot (default: "Linear Regression Model Analysis")}
}
\value{
models list, CV results for k=3/5/LOO and a plot of choice. interactive.
}
\description{
Screen for models using parallel processing, cross validate them and plot. Designed for interactive work.
}
