% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_models.R
\name{model.report}
\alias{model.report}
\title{Generate models and a plot}
\usage{
model.report(
  dataset,
  min = 2,
  max = floor(dim(mod_data)[1]/5),
  leave.out = "",
  predict = F
)
}
\arguments{
\item{dataset}{a dataframe with outcome column (must be named 'output')}

\item{min}{minimum # of features (default = 2)}

\item{max}{max # of features (defaults = # of observations / 5)}

\item{leave.out}{name of observations to leave out (e.g. 'p_Br')}

\item{predict}{if leave.out is not empty, should a prediction
of it be computed?}
}
\value{
models list, CV results for k=3/5/LOO and a plot of choice. interactive.
}
\description{
Screen for models, cross validate them and plot. Designed for interactive work.
}
