% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parallel_linear_models.R
\name{model.single.cv.iterator}
\alias{model.single.cv.iterator}
\title{Support funcion - iterator for CV}
\usage{
model.single.cv.iterator(i)
}
\arguments{
\item{i}{iteration number}
}
\value{
iteration MAE and Q2
}
\description{
Support funcion - iterator for CV
}
\keyword{internal}
