% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_models.R
\name{model.single.cv}
\alias{model.single.cv}
\title{Cross validate (k-fold) a single model}
\usage{
model.single.cv(formula, data, out.col, folds)
}
\arguments{
\item{formula}{model formula (not designed for direct use)}

\item{data}{data frame with outcome column}

\item{out.col}{column number with output}

\item{folds}{define K}
}
\value{
MAE and Q2
}
\description{
A utility function for the brute force search of models
}
\keyword{internal}
