% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logistic_models.R
\name{model.subset.log.ordinal}
\alias{model.subset.log.ordinal}
\title{Get ordinal logistic models of all feature subsets within a scope}
\usage{
model.subset.log.ordinal(
  data,
  out.col = which(colnames(data) == "class"),
  min = 1,
  max = floor(dim(data)[1]/5)
)
}
\arguments{
\item{data}{a data frame with a class column}

\item{out.col}{what column holds the class}

\item{min}{minimum number of variables}

\item{max}{maximum number of variables}
}
\value{
list of models, ranked by McFadden's pseudo R squared
}
\description{
Get ordinal logistic models of all feature subsets within a scope
}
