% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parallel_linear_models.R
\name{model.subset.parallel}
\alias{model.subset.parallel}
\title{Brute force model search with a min and max number of
features. A parallel computation version.}
\usage{
model.subset.parallel(
  data,
  out.col = dim(data)[2],
  min = 2,
  max = floor(dim(data)[1]/5),
  folds = nrow(data),
  iterations = 1,
  cutoff = 0.85,
  cross.terms = F
)
}
\arguments{
\item{data}{data frame with output column}

\item{out.col}{number of output column}

\item{min}{minimum # of features (default = 2)}

\item{max}{max # of features (defaults = # of observations / 5)}

\item{folds}{defaults to nrow(data)}

\item{iterations}{defaults to 1 (LOOCV)}

\item{cutoff}{search for Q2 above 0.85 (if there isn't will look for lower)}

\item{cross.terms}{if TRUE includes feature interactions (explosive - try avoiding)}
}
\value{
table with 10 best models (at max)
}
\description{
Ranks models based on K-fold CV Q2
}
