% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parallel_linear_models.R
\name{models.list.parallel}
\alias{models.list.parallel}
\title{Generate a .csv file with top 10 linear models resulting from a parallel 
 model.subset search.}
\usage{
models.list.parallel(
  dataset,
  min = 2,
  max = floor(dim(mod_data)[1]/5),
  leave.out = "",
  folds = nrow(mod_data),
  iterations = 1
)
}
\arguments{
\item{dataset}{a dataframe with outcome column (must be named 'output')}

\item{min}{minimum # of features (default = 2)}

\item{max}{max # of features (defaults = # of observations / 5)}

\item{leave.out}{name of observations to leave out (e.g. 'p_Br')}

\item{folds}{number of folds for CV}

\item{iterations}{number of iterations for CV}
}
\value{
models list, CV results for k=3/5/LOO.
}
\description{
Screen for models and cross validate them.
}
