% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_angles_bonds.R
\name{mol.angles}
\alias{mol.angles}
\title{Compute angles and dihedrals}
\usage{
mol.angles(atoms)
}
\arguments{
\item{atoms}{a character}
}
\value{
1x1 data frame of the angle/dihedral value in degrees
}
\description{
This function acts on a read feather file
Input is a character of either 3 or 4 atoms with
atoms separated by a space (e.g. '1 2 3' or '1 2 3 4')
}
\keyword{internal}
