% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_angles_bonds.R
\name{mol.angles.set}
\alias{mol.angles.set}
\title{Compute several angles and dihedrals in a single molecule}
\usage{
mol.angles.set(atoms_vector)
}
\arguments{
\item{atoms_vector}{a vector of characters}
}
\value{
1x(num of sets) data frame of the angle/dihedral values in degrees
}
\description{
This function acts on a read feather file
Input is a character of either 3 or 4 atoms with
atoms separated by a space (e.g. '1 2 3' or '1 2 3 4')
and sets gathered in a vector (e.g. c('1 2 3', '1 2 3 4'))
}
\keyword{internal}
