% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R, R/moleculaR-package.R
\docType{package}
\name{moleculaR-package}
\alias{moleculaR-package}
\alias{_PACKAGE}
\alias{imports-package}
\alias{import_usethis}
\title{moleculaR: moleculaR - Extract and Compute User Guided Quantum-Chemical Molecular Features}
\description{
A physical-organic chemistry toolset - designed for easy, chemical-intuition based molecular features extarction and statistical modeling.

A physical-organic chemistry toolset - designed for easy, chemical-intuition based molecular features extarction and statistical modeling.
}
\author{
\strong{Maintainer}: Shahar Barkai \email{barkais@post.bgu.ac.il} (\href{https://orcid.org/0000-0001-7242-6551}{ORCID})

}
\keyword{internal}
