% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moleculaR_main.R
\name{moleculaR.input}
\alias{moleculaR.input}
\title{User function for the extraction of all possible features
based on a ready input file}
\usage{
moleculaR.input(input_file = NULL)
}
\arguments{
\item{input_file}{moleculaR's input file}
}
\value{
csv file with all features and a input file for future
 use and documentation
}
\description{
Calculate all features by applying to an input file, saving the final
 data set in a chosen location.
}
