% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moleculaR.R
\docType{package}
\name{moleculaR}
\alias{moleculaR}
\title{moleculaR: Spatial Probabilistic Mapping of Metabolite Ensembles in Mass Spectrometry Imaging}
\section{Description}{


The moleculaR package provides a computational framework that introduces probabilistic mapping
and point-for-point statistical testing of metabolites in tissue via Mass spectrometry imaging.
It enables collective molecular projections and consequently spatially-resolved investigation
of ion milieus, lipid pathways or user-defined biomolecular ensembles within the same image.
}

\section{Details}{


Package: \tab moleculaR \cr
License: \tab GPL (>= 3) \cr
URL: \tab  \cr
}

\section{Main Classes}{

}

\section{Authors}{

}

\section{References}{

}

