% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{molProbMap}
\alias{molProbMap}
\title{molProbMap Class Constructor}
\usage{
molProbMap(
  bw,
  sppMoi,
  csrMoi,
  rhoMoi,
  rhoCsr,
  hotspotpp,
  hotspotIm,
  hotspotMask,
  coldspotpp,
  coldspotIm,
  coldspotMask
)
}
\arguments{
\item{bw}{A numeric, the calculated Gaussian bandwidth}

\item{sppMoi}{A 'analytePointPattern' object representing a given analyte (metabolite-of-interest).}

\item{csrMoi}{A 'analytePointPattern' object representing the corresponding CSR model.}

\item{rhoMoi}{An 'im' object storing the density image of 'sppMoi' after applying KDE.}

\item{rhoCsr}{An 'im' object storing the density image of 'csrMoi' after applying KDE.}

\item{hotspotpp}{A 'analytePointPattern' object holding only the points that are located within the hotspot.}

\item{hotspotIm}{An 'im' object storing the corresponding pixellated image of the 'hotspotpp' slot.}

\item{hotspotMask}{An 'owin' object storing the corresponding window of the 'hotspotpp' slot.}

\item{coldspotpp}{A 'analytePointPattern' object holding only the points that are located within the coldspot.}

\item{coldspotIm}{An 'im' object storing the corresponding pixellated image of the 'coldspotpp' slot.}

\item{coldspotMask}{An 'owin' object storing the corresponding window of the 'coldspotpp' slot.}
}
\value{
An S3 object of type 'molProbMap'.
}
\description{
This is an S3 class to represent the calculated molecular probability maps. This
class is only used with objects returned by `porbMap` function and should not be
called explicitly by the user.
}
\keyword{internal}
