% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping-functions.R
\name{mse}
\alias{mse}
\title{Mean Squared Error}
\usage{
mse(im0, im1)
}
\arguments{
\item{im0:}{a reference image object of type `im`.}

\item{im1:}{a second image object of type `im`.}
}
\value{
A numeric, the calculated mean squared error.
}
\description{
This function Calculates the mean squared error
 between two
images of type `im` (see `spatstat` documentation).
}
\keyword{internal}
