% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{name.vib}
\alias{name.vib}
\title{helper function for naming}
\usage{
name.vib(x, atom_pairs)
}
\arguments{
\item{x}{bond number}

\item{atom_pairs}{bonds}
}
\value{
names
}
\description{
helper function for naming
}
\keyword{internal}
