% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nbos.R
\name{nbo.df}
\alias{nbo.df}
\title{Pull NBO charges for specific atoms}
\usage{
nbo.df(atom_indices, difference_indices = NA)
}
\arguments{
\item{atom_indices}{a character of atoms}

\item{difference_indices}{a character of atom pairs}
}
\value{
A data frame with NBO charges and differences
}
\description{
Choose atoms for which you wish to have NBO charges presented.
Charge differences are also available.
}
