% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_charges.R
\name{nbo.info}
\alias{nbo.info}
\title{Pull NBO charges for specific atoms}
\usage{
nbo.info(atom_index)
}
\arguments{
\item{atom_index}{a character of atoms}
}
\value{
A data frame with NBO charges
}
\description{
Choose atoms for which you wish to have NBO charges presented
}
\keyword{internal}
