% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{ncc}
\alias{ncc}
\title{Normalized cross correlation}
\usage{
ncc(im0, im1)
}
\arguments{
\item{im0:}{a reference image object of type `im`.}

\item{im1:}{a second image object of type `im`.}
}
\value{
A numeric, the calculated normalized cross correlation
}
\description{
This function Calculates the normalized cross correlation between two
images of type `im` (see `spatstat` documentation).
}
\keyword{internal}
