% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz_steRimol.R
\name{NOB.Atype}
\alias{NOB.Atype}
\title{determines atom type based on number of bonds}
\usage{
NOB.Atype(row, substi, bonds)
}
\arguments{
\item{row}{row number in substi (atom)}

\item{substi}{truncated xyz}

\item{bonds}{bonds data frame}
}
\value{
list of atom types
}
\description{
determines atom type based on number of bonds
}
\keyword{internal}
