% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{npa.dipole.subunit.df}
\alias{npa.dipole.subunit.df}
\title{Extract NBO charge based DM of a substructure alone}
\usage{
npa.dipole.subunit.df(coor_atoms = NA, sub_atoms = NA)
}
\arguments{
\item{coor_atoms}{3 atoms character (e.g. '1 2 3')}

\item{sub_atoms}{atoms character (e.g. '4 5 6 7 8 9')}
}
\value{
dataframe with npa based DM of a substructure
}
\description{
Extract NBO charge based DM of a substructure alone
}
\keyword{internal}
