% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{npa.dipole.subunit.multi}
\alias{npa.dipole.subunit.multi}
\title{Compute NBO based DM for several substructures}
\usage{
npa.dipole.subunit.multi(coor_atoms, subunits_inputs_vector)
}
\arguments{
\item{coor_atoms}{3 atoms character (e.g. '1 2 3')}

\item{subunits_inputs_vector}{vector of atoms characters
(e.g. c('10,11,12,13', '4 5 6 7 8 9'))}
}
\value{
dataframe with npa based DM of a substructure
}
\description{
Compute NBO based DM for several substructures
}
