% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{npa.dipole.subunit}
\alias{npa.dipole.subunit}
\title{Compute nbo based dipole moment for a subset of atoms}
\usage{
npa.dipole.subunit(coor_atoms, subunit)
}
\arguments{
\item{coor_atoms}{3 atoms character (e.g. '1 2 3')}

\item{subunit}{atoms character (e.g. '4 5 6 7 8 9')}
}
\value{
npa dipole moment for a subsunit
}
\description{
Compute nbo based dipole moment for a subset of atoms
}
\keyword{internal}
