% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{nssd}
\alias{nssd}
\title{Normalized sum of squared differences}
\usage{
nssd(im0, im1)
}
\arguments{
\item{im0:}{a reference image object of type `im`.}

\item{im1:}{a second image object of type `im`.}
}
\value{
A numeric, the calculated normalized sum of squared differences.
}
\description{
This function Calculates the normalized sum of squared differences between two
images of type `im` (see `spatstat` documentation).
}
\keyword{internal}
