% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moleculaR_main.R
\name{parse_txt_inputfile_to_list}
\alias{parse_txt_inputfile_to_list}
\title{Text Input file parser to moleculaR's main function}
\usage{
parse_txt_inputfile_to_list(input_file_path)
}
\value{
list form of input file
}
\description{
Transform into list form.
}
\keyword{internal}
