% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{plot.molProbMap}
\alias{plot.molProbMap}
\title{Plot molecular Probability Maps}
\usage{
\method{plot}{molProbMap}(
  obj,
  what = "detailed",
  transpFactor = 0.7,
  signifArea = "both",
  analyte = "m/z Analyte",
  ionImage = NA
)
}
\arguments{
\item{obj}{S3 object of type `molProbMap`.}

\item{what}{What to plot, c("detailed", "sppMoi", "csrPointPattern", )}

\item{transpFactor}{Transparency fraction. Numerical value or
vector of values between 0 and 1, giving the opaqueness of a colour.
A fully opaque colour has `transpFactor=1`.}

\item{signifArea}{a character indicating which significance area to plot i.e. c("both", "hotspot", "coldspot").}

\item{analyte}{character, name of the analyte.}

\item{ionImage}{an optional rastered image of type `spatstat::im` of the corresponding "regular"
ion image, used for comparison. Could be generated via `moleculaR::searchAnalyte(..., wMethod = "sum")`
and subsequently using `spatstat.geom::pixellate`.}
}
\value{
nothing, plots only.
}
\description{
A method to plot MPMs with different combinations.
}
