% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecule_plot.R
\name{plot_molecule}
\alias{plot_molecule}
\title{Generate a 3D plot of an xyz file, including atom indices}
\usage{
plot_molecule(xyz_file)
}
\arguments{
\item{xyz_file}{an xyz file}
}
\value{
plot widget (Mac and Windows only, rgl is a mess on linux)
}
\description{
Generate a 3D plot of an xyz file, including atom indices
}
