% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{plotAnalyte}
\alias{plotAnalyte}
\title{Plot analytePointPattern}
\usage{
plotAnalyte(
  obj,
  colourPal = "inferno",
  uniformCol = NULL,
  transpFactor = 0.7,
  rescale = TRUE,
  pch = 19,
  size = 0.4,
  main = NULL,
  leg.side = "right",
  leg.args = list(cex = 3, cex.axis = 1.25),
  ...
)
}
\arguments{
\item{obj}{S3 object of type `anlaytePointPattern` (and `spp`).}

\item{colourPal}{the colourmap to be used, see `?viridis::viridis_pal`.}

\item{uniformCol}{a character specifying a single colour. This will override `colourPal`.}

\item{transpFactor}{Transparency fraction. Numerical value or vector of values between 0 and 1,
giving the opaqueness of a colour. A fully opaque colour has `transpFactor=1`.}

\item{rescale}{logical, whether to scale the intensities of the output plot
to [0,1] interval.}

\item{pch}{a positive integer, or a single character. See `?par`.}

\item{size}{the size of the symbol: a positive number or zero. see`?symbolmap`.}

\item{main}{character, title of the plot. If not given, the m/z value of `obj` is used.}

\item{leg.side}{position of legend relative to main plot.}

\item{leg.args}{list of additional arguments passed to control the legend. see
`?spatstat.geom::plot.ppp` for more details.}

\item{...}{further arguments passed to `spatstat.geom::plot.ppp`.}
}
\value{
nothing, plots only.
}
\description{
A method to plot `anlaytePointPattern` objects. This calls `plot.ppp` with
pre-defined defaults. For more control over the plotting please use `plot.ppp`.
}
