% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{plotImg}
\alias{plotImg}
\title{Plot raster images of type `im`}
\usage{
plotImg(
  obj,
  colourPal = "inferno",
  uniformCol = NULL,
  rescale = TRUE,
  transpFactor = 0.7,
  irange = NULL,
  ribargs = list(las = 2, cex.axis = 1.25),
  ribsep = 0.05,
  ...
)
}
\arguments{
\item{obj}{S3 object of type `im`.}

\item{colourPal}{the colourmap to be used, see `?viridis::viridis_pal`.}

\item{uniformCol}{a character specifying a single colour. This will override `colourPal`.}

\item{rescale}{logical, whether to scale the intensities of the output plot
to [0,1] interval.}

\item{transpFactor}{Transparency fraction. Numerical value or vector of values between 0 and 1,
giving the opaqueness of a colour. A fully opaque colour has `transpFactor=1`.}

\item{irange}{a numeric of length 2, a custome intensity range for plotting. Incompatible with `rescale`;
if provided, `rescale` will be set to `FALSE`.}

\item{ribargs}{a list of additional arguments to control the display of the ribbon. see
`?spatstat.geom::plot.im` for details.}

\item{ribsep}{a numeric controlling the space between the ribbon and the image.}

\item{...}{further arguments passed to `plot.im`.}
}
\value{
nothing, plots only.
}
\description{
A method to plot `im` objects of the `spatstat` package. This calls `plot.im` with
pre-defined defaults. For more control over the plotting please use `plot.im`.
}
