% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polar.R
\name{polar.df}
\alias{polar.df}
\title{Pull polarizability info for a set of molecules}
\usage{
polar.df()
}
\value{
A data frame with polarizabilities
}
\description{
Pull polarizability info for a set of molecules
}
