% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logistic_models.R
\name{prob.heatmap}
\alias{prob.heatmap}
\title{Create a nice probabilities of predictions table}
\usage{
prob.heatmap(model, data)
}
\arguments{
\item{model}{a classifier model (nnet/MASS)}

\item{data}{a data frame with a class column}
}
\value{
probabilities table figure
}
\description{
Create a nice probabilities of predictions table
}
