% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{probMap}
\alias{probMap}
\title{Calculate molecular probablity maps}
\usage{
probMap(
  spp,
  win,
  weighted = TRUE,
  control = NULL,
  bw = NULL,
  bwMethod = "spAutoCor",
  dists = seq(1, 5, 0.5),
  numSim = 99,
  approximate = FALSE,
  approxRadius = NULL,
  csrIntensities = "resample",
  csr = NULL,
  bwPlot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{spp:}{The spatial point pattern.}

\item{win:}{The window object of type `spatstat::owin`.}

\item{weighted:}{Whether the intensities are inlcude into the computation.}

\item{bw:}{The Gaussian bandwidth used for Kernel density estimation.}

\item{bwMethod:}{The method used for computing the Gaussian bandwidth, c("spAutoCor","iterative", "scott").}

\item{dists:}{for `bwMethod="spAutoCor"`, these are neighborhood distances at which to compute the Moran's I statisitc.
for `bwMethod="iterative"` these are the Gaussian bandwidth upon which to iterate.}

\item{numSim:}{the number of Monte Carlo simulations to be done for Moran's I, see ?spdep::moran.mc.
(only relevant with `bwMethod="spAutoCor"`).}

\item{approximate:}{for `bwMethod="spAutoCor"`, when `TRUE`only points within a circle centered at the intensity-wighted
centroid (center of gravity) and radius `approxRadius` are considered. This is used to speed up the caluclation.}

\item{approxRadius:}{for `bwMethod="spAutoCor"`, the radius in units (usually pixels) for the approximation circlular region.By default,
this equates to the radius that produces a circular area which is half of the area of the spp window.}

\item{bwPlot:}{whether to plot the Gaussian bw selection procedure, ignored when \code{bwMethod = "scott"}.}

\item{csrIntensities:}{How the intensities for the CSR model are generated, c("resample", "Poisson", "Gaussian").}

\item{control:}{An spp object that is designated as the "control" or "null" alternative of \code{spp}. If supplied the
CSR model will be generated from the intensities of this object.}

\item{csr:}{Pre-computed wighted csr model corresponding to the given spp. This
could be useful when generating generating hotspot-bw curves.}

\item{verbose:}{whether to output progress.}
}
\value{
A list of
cutoff: the chosen threshold above which it is a hotspot,
GausBW: calculated Gaussian bandwidth,
csrspp: calculated complete spatial randomness model for the input spp,
denspp: density image of the input point pattern spp,
denCsr: density image of the computed csr pattern,
hotspotpp: the remaining points which lie within the hotspot mask as a spatial point pattern,
hotspotIm: the hotspot image,
hotspotMask: the hotspot mask of type owin.mask,
nonHotspotMask: the non-hotspot areas, to be used to mask the non hotspot areas.
}
\description{
This function creates molecular probability maps from a given spatial point pattern representation.
}
