% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Radii.Pyykko}
\alias{Radii.Pyykko}
\title{Radii.Pyykko}
\description{
A data frame with covalent radii values, as they are given in
dx.doi.org/10.1021/jp5065819 | J. Phys. Chem. A 2015, 119, 2326−2337
}
\keyword{internal}
