% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCentrData.R
\name{readCentrData}
\alias{readCentrData}
\title{Read Centroided imzML}
\usage{
readCentrData(path, verbose = FALSE)
}
\arguments{
\item{path:}{character string, path to the imzML file which should be read in.}

\item{verbose:}{logical, verbose output?}
}
\value{
A list of \code{MALDIquant::MassPeaks} objects.
}
\description{
A wrapper for \code{MALDIquantForeign::importImzMl} for reading centroided (i.e. processed)
imzML data.
}
