% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSingleSpect.R
\name{readSingleSpect}
\alias{readSingleSpect}
\title{Read a single spectrum}
\usage{
readSingleSpect(path)
}
\arguments{
\item{path:}{character string, path to the tsv file which should be read in.}
}
\value{
A \code{MALDIquant::MassSpectrum} object.
}
\description{
This reads a single spectrum from a tsv file containing two columnd for m/z and intensity,
respectively.
}
