% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{ring.vib.conjugated}
\alias{ring.vib.conjugated}
\title{Get coupled rings characteristic vibrations' frequencies}
\usage{
ring.vib.conjugated(primary_atom_1, primary_atom_2)
}
\arguments{
\item{primary_atom_1}{The atom index of the first ring atom bonded to the basic
strcuture.}

\item{primary_atom_2}{The atom index of the second ring atom bonded to the basic
strcuture.}
}
\value{
A data frame with stretching frequencies
}
\description{
Get coupled rings characteristic vibrations' frequencies
}
\keyword{internal}
