% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{ring.vib.df}
\alias{ring.vib.df}
\title{Get a ring's characteristic vibrations frequencies, for a set of molecules}
\usage{
ring.vib.df(input = NA)
}
\arguments{
\item{input}{a character of 6 atoms, ordered as follows:
para (opposite of primary - doesn't matter, but must be consistent),
two ortho atoms to primary and two meta atoms to primary.
 See full manual for a detailed explanation.}
}
\value{
A data frame with stretching frequencies
}
\description{
Get a ring's characteristic vibrations frequencies, for a set of molecules
}
