% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{ring.vib}
\alias{ring.vib}
\title{Get a ring's characteristic vibrations frequencies}
\usage{
ring.vib(ordered_ring_atoms)
}
\arguments{
\item{ordered_ring_atoms}{a character of 6 atoms, ordered as follows:
para (opposite of primary - doesn't matter, but must be consistent),
two ortho atoms to primary and two meta atoms to primary.
 See full manual for a detailed explanation.}
}
\value{
A data frame with stretching frequencies
}
\description{
Get a ring's characteristic vibrations frequencies
}
\keyword{internal}
