% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Runs one of the campanion shiny apps}
\usage{
runShiny(app)
}
\arguments{
\item{app:}{a character, name of the shiny app, \code{c("package-app", "example-app")}.}
}
\value{
Has no return. To quit simply interrupt R by pressing Ctrl+c or Esc.
}
\description{
Calls one of the shiny apps accompanying this package. \code{package-app} is the
workhorse app of the \code{moleculaR}. \code{example-app} only load an filtered
example dataset and is used only for showcasing the functionalities.
}
