% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAnalyte.R
\name{searchAnalyte}
\alias{searchAnalyte}
\title{Analyte detection in an MSI dataset}
\usage{
searchAnalyte(
  m,
  fwhm,
  spData,
  wMethod = "Gaussian",
  verifiedMasses = NA,
  confirmedOnly = FALSE,
  metaData = list()
)
}
\arguments{
\item{m}{the mass to be queried (in Da).}

\item{fwhm}{the fwhm at 'm'.}

\item{spData}{an S3 object of type 'sparseIntensityMatrix' holding the sparse MSI data.}

\item{wMethod}{wighting method; c("Gaussian", "sum", "max", "mean").}

\item{verifiedMasses}{an optional numeric vector of m/z values that are (externally) verified to
be real molecular entities with a certain confidence, ex. 'mz' column of a METASPACE
annotation result.}

\item{confirmedOnly}{if `TRUE`, returns detections only if confirmed by `verifiedMasses`.}

\item{metaData}{optional named list with additional identifiers for the analyte
under study, ex. list(lipidID = "..", sumformular = "..", ..). This will be passed to the `metaData`
slot of the resulting `analytePointPattern` object.}
}
\value{
An analyte point patter of type `ppp` and `analytePointPattern` containing the hits for
`m` in `msData`. If `confirmedOnly = TRUE` then only these hits are returned.
}
\description{
This function check if an analyte (given by m/z) is detectable in the MS data represented
by a 'sparseIntensityMatrix' object.
}
