% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchLipid.R
\name{searchLipid}
\alias{searchLipid}
\title{Find lipids in MSI datasets}
\usage{
searchLipid(
  m,
  fwhm,
  massAxis,
  spData,
  coords,
  mtspc = NA,
  adduct = NA,
  mode = NA,
  modeAdduct = NA,
  lipidID = NA,
  sumformula = NA,
  abbrev = NA,
  numDoubleBonds = NA
)
}
\arguments{
\item{sumformula}{for internal use only.}

\item{abbrev}{for internal use only.}

\item{numDoubleBonds}{for internal use only.}

\item{m:}{the mass to be queried (Da).}

\item{fwhm:}{the fwhm at `m`.}

\item{massAxis:}{the mass axis of the whole MS dataset to be searched against.}

\item{spData:}{the sparse matrix containing the negative or poisitive ion mode info.}

\item{coords:}{the coordinates of spectra contained within the spData.}

\item{mtspc:}{a data frame of the metaspcae identifications.}

\item{lipidID:}{for internal use only.}

\item{fullName}{for internal use only.}
}
\value{
A dataframe containing the hits for `m` in `msData`.
}
\description{
This function check if a deprotonated/protonated or any alkali version lipid is detectable in the MS data represented
by a sparse matrix.
}
