% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkData.R
\name{shrinkData}
\alias{shrinkData}
\title{Shrink data}
\usage{
shrinkData(x, mzKeep, fwhmFun)
}
\arguments{
\item{x:}{Dataset, a list of \code{MALDIquant::MassPeaks} objects.}

\item{mzKeep:}{numeric vector, m/z values to retain.}

\item{fwhmFun:}{a function, fwhm as a function of m/s axis. See \code{?moleculaR::estimateFwhm}.}
}
\value{
Filtered list of \code{MALDIquant::MassPeaks} objects.
}
\description{
This function is for internal use only. It tries to shrink MSI data (in terms of size) for including it within
a package.
}
\keyword{internal}
