% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logistic_models.R
\name{simi.sampler}
\alias{simi.sampler}
\title{a similarity based sampling algorithm}
\usage{
simi.sampler(
  data,
  class,
  compare.with = 0,
  plot = F,
  sample.size = min(summary(as.factor(data$class)))
)
}
\arguments{
\item{data}{data frame with class column}

\item{class}{class of interest number}

\item{compare.with}{choose the class with which similarity is computed.
Defaults to 0, being similarity of each sample with its own group. 
Any other number will compare with the class represented by that number.}

\item{plot}{create a plot of the similarity before and after sampling.}

\item{sample.size}{how many are to be sampled. 
Defaults to the number of sample in the smallest class.}
}
\value{
A vector with row numbers of sampled observations
}
\description{
Used to resize (under-sample) classification oriented data sets.
Returns the samples row numbers.
}
