% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-functions.R
\name{simulatespp}
\alias{simulatespp}
\title{Simulate a Spatial Point Pattern}
\usage{
simulatespp(
  lambdaHotspot,
  lambdaBackground,
  intensityData,
  radius,
  backgroundUpperThr = 0.75,
  hotspotLowerThr = 0.75,
  shape = "centralCircle",
  backgroundLength = 100,
  seed = NULL,
  gtPlot = FALSE,
  sppPlot = TRUE
)
}
\description{
-- under development --
}
