% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{sparseIntensityMatrix}
\alias{sparseIntensityMatrix}
\title{sparseIntensityMatrix Class Constructor}
\usage{
sparseIntensityMatrix(spmat, mzAxis, coordinates)
}
\arguments{
\item{spmat}{A sparse matrix of type 'Matrix::dgCMatrix' holding MSI intensity data.}

\item{mzAxis}{A numeric, the overall m/z axis for the experiment.}

\item{coordinates}{A two-column matrix, the coordinates of the spectra which are stored
in rows of 'spmat'.}
}
\value{
An S3 object 'sparseIntensityMatrix'.
}
\description{
An S3 class to hold mass spectrometry imaging (MSI) intensity data in a sparse representation ('Matrix::dgCMatrix' class)
with few additional slots. It is used to speed-up computation.
}
