% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helping-functions.R
\name{spp2im}
\alias{spp2im}
\title{Convert `anlaytePointPattern` to `im`}
\usage{
spp2im(obj, weighted = TRUE, rescale = FALSE, zero.rm = FALSE, ...)
}
\arguments{
\item{obj}{S3 object of type `anlaytePointPattern` (and `spp`).}

\item{weighted}{a logical, whether to scale points by their intensities.}

\item{rescale}{logical, whether to scale the intensities of the output `im` object
to [0,1] interval.}

\item{zero.rm}{for internal use only.}

\item{...}{additional arguments passed to `spatstat.geom::pixellate`.}
}
\value{
an object of type `im` (see `spatstat` documentation).
}
\description{
A method to convert `anlaytePointPattern` objects to `im` objects.
}
\keyword{internal}
