% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cube_steRimol.R
\name{steRimol.cube.multi}
\alias{steRimol.cube.multi}
\title{Calculate cube based sterimol parameters for multiple coordinate sets}
\usage{
steRimol.cube.multi(
  coordinates_vector,
  only.sub = T,
  drop = NULL,
  degrees = 90,
  isovalue = 0.036
)
}
\arguments{
\item{coordinates_vector}{Vector of coordinate strings (e.g. c("1 2", "3 4"))}

\item{only.sub}{If TRUE (default) will account only for atoms directly bonded}

\item{drop}{Numeric value of atoms to drop from calculation}

\item{degrees}{Define rotational scan, defaults to 90}

\item{isovalue}{Minimum density isovalue to consider, defaults to 0.036}
}
\value{
Data frame with sterimol values for each coordinate set
}
\description{
Calculate cube based sterimol parameters for multiple coordinate sets
}
