% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz_steRimol.R
\name{steRimol}
\alias{steRimol}
\title{Verloop's sterimol values from moleculaR's csv files}
\usage{
steRimol(coordinates, CPK = T, only_sub = T, drop = NULL, plot.B1 = T)
}
\arguments{
\item{coordinates}{primary axis, a two atom character}

\item{CPK}{logical, if TRUE will use CPK radii based on Verloop's atom type,
if FALSE will use Pyykko's covalent radii}

\item{only_sub}{if TRUE (default) will account only for atoms directly bonded
from the primary axis onward}

\item{drop}{numeric value of an atom, from which onward atoms will be dropped from calculation}

\item{plot.B1}{should a plot of B1 and B5 box be presented}
}
\value{
sterimol values for a single molecule
}
\description{
Calculate a single molecule's sterimol values
}
\keyword{internal}
