% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz_steRimol.R
\name{steRimol.xyz.df}
\alias{steRimol.xyz.df}
\title{Verloop's sterimol values from xyz files}
\usage{
steRimol.xyz.df(coordinates, CPK = T, only_sub = T, drop = NULL)
}
\arguments{
\item{coordinates}{primary axis, a two atom character}

\item{CPK}{logical, if TRUE will use CPK radii based on Verloop's atom type,
if FALSE will use Pyykko's covalent radii}

\item{only_sub}{if TRUE (default) will account only for atoms directly bonded
from the primary axis onward}

\item{drop}{numeric value of an atom, from which onward atoms will be dropped from calculation}
}
\value{
sterimol values for a set of molecules
}
\description{
Calculate a set of molecules' sterimol values
}
