% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz_steRimol_parallel.R
\name{steRimol.xyz.multi.parallel}
\alias{steRimol.xyz.multi.parallel}
\title{Verloop's sterimol values from xyz files - parallel version (for Mac and Linux)}
\usage{
steRimol.xyz.multi.parallel(
  coordinates_vector,
  CPK = T,
  only_sub = T,
  drop = NULL
)
}
\arguments{
\item{coordinates_vector}{primary axes, a vector of two atom characters}

\item{CPK}{logical, if TRUE will use CPK radii based on Verloop's atom type,
if FALSE will use Pyykko's covalent radii}

\item{only_sub}{if TRUE (default) will account only for atoms directly bonded
from the primary axis onward}

\item{drop}{numeric value of an atom, from which onward atoms will be dropped from calculation}
}
\value{
sterimol values (multiple)
}
\description{
Calculate multiple sterimol values for a set of molecules
}
