% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz_steRimol_parallel.R
\name{steRimol.xyz.parallel}
\alias{steRimol.xyz.parallel}
\title{Verloop's Sterimol Values from XYZ Files - - parallel version (for Mac and Linux)}
\usage{
steRimol.xyz.parallel(
  mol,
  coordinates,
  CPK = T,
  only_sub = T,
  drop = NULL,
  plot.B1 = T
)
}
\arguments{
\item{mol}{A data frame representing the molecular structure from an XYZ file.}

\item{coordinates}{A character string specifying the primary axis, defined by two atom indices separated by a space.}

\item{CPK}{Logical; if TRUE (default), uses CPK radii based on Verloop's atom types; if FALSE, uses Pyykkö's covalent radii.}

\item{only_sub}{Logical; if TRUE (default), only considers atoms directly bonded from the primary axis onward.}

\item{drop}{Numeric vector; specifies atom indices to exclude from calculations.}

\item{plot.B1}{Logical; if TRUE, generates a plot of the B1 and B5 box.}
}
\value{
A data frame containing Sterimol values (B1, B5, L) and related parameters for the given molecule.
}
\description{
Calculates Sterimol parameters (B1, B5, L) for a given molecule based on atomic coordinates.
}
\keyword{internal}
