% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_logistic_models.R
\name{stratified}
\alias{stratified}
\title{Stratify data sets by a group}
\usage{
stratified(df, group, size)
}
\arguments{
\item{df}{a dataframe with a categorial column}

\item{group}{group by which stratified}

\item{size}{a numeric above or below 1, with above sampling the number
and below sampling the ratio.}
}
\value{
a stratum df
}
\description{
User defined stratification by a class column. 
Designated for k-fold CV.
}
\keyword{internal}
