% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{stretch.vib.df}
\alias{stretch.vib.df}
\title{Get a bond's stretching frequency for a set of molecules}
\usage{
stretch.vib.df(atom.pairs, threshold = 1350)
}
\arguments{
\item{atom.pairs}{a character of atom pairs}

\item{threshold}{minimal frequency value (above fingerprint region)}
}
\value{
A data frame with stretching frequencies
}
\description{
Get a bond's stretching frequency for a set of molecules
}
