% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{stretch.vib}
\alias{stretch.vib}
\title{Get a bond's stretching frequency}
\usage{
stretch.vib(atom_pairs, threshold = 1350)
}
\arguments{
\item{atom_pairs}{a character of atom pairs}

\item{threshold}{minimal frequency value (above fingerprint region)}
}
\value{
A data frame with stretching frequencies
}
\description{
Get a bond's stretching frequency
}
\keyword{internal}
