% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetAnalytes.R
\name{subsetAnalytes}
\alias{subsetAnalytes}
\title{Subset analytePointPatterns}
\usage{
subsetAnalytes(obj, expr)
}
\arguments{
\item{obj}{an S3 object of type `ppp` and `analytePointPatterns`.}

\item{expr}{logical expression indicating which points are to be kept. The expression must involve
the column names of the `metaData` slot). Subsetting based on `marks` is not yet supported.}
}
\value{
An analyte point patter of type `ppp` and `analytePointPattern` after subsetting.
}
\description{
subset a given point patterns of type `ppp` and `analytePointPatterns` based on its `metaData`.
}
