% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-superimposeNoise.R, R/superimposeNoise.R
\name{superimposeNoise}
\alias{superimposeNoise}
\title{Superimpose noise over MSI data - only for testing}
\usage{
superimposeNoise(
  x = NULL,
  spData = NULL,
  method,
  mz,
  fwhm,
  noiseFactor = 1L,
  searchFactor = 3L,
  sigmaInterfering = 2L,
  mzTrim = 0,
  numSpikedPeaks = 10L,
  returnMat = FALSE,
  verbose = FALSE
)

superimposeNoise(
  x = NULL,
  spData = NULL,
  method,
  mz,
  fwhm,
  noiseFactor = 1L,
  searchFactor = 3L,
  sigmaInterfering = 2L,
  mzTrim = 0,
  numSpikedPeaks = 10L,
  returnMat = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x:}{dataset, a list of \code{MALDIquant::MassPeaks} objects.}

\item{spData:}{a corresponding sparse MSI matrix of type 'sparseIntensityMatrix'.}

\item{method:}{character string, the method used to add noise c("Poisson", "Gaussian","intensityArtifacts","interfering").}

\item{mz:}{numeric, m/z value where the noise will be added.}

\item{fwhm:}{numeric, the fwhm at the specified m/z.}

\item{noiseFactor:}{integer, this will be multiplied by the standard deviation of intensities at \code{mz} which will
be plugged into \code{rnorm} to generate the intensities of the Gaussian noise.}

\item{searchFactor:}{integer, search tolerance, how many standard deviations (sigma) around \code{mz} to include in the search of
\code{mz} peak presence.}

\item{sigmaInterfering:}{integer, how many standard deviations (sigma) away from \code{mz} to place the interfering peak.}

\item{mzTrim:}{numeric, this will trim the m/z axis to only include \code{[mz-mzTrim,mz+mzTrim]}. Skipped if set to \code{0}
(default) or \code{length(mz) > 1}.}

\item{numSpikedPeaks:}{integer, number of spiked peaks for \code{intensityArtifacts} method.}

\item{returnMat:}{if \code{TRUE} returns a sparse MSI matrix of type `sparseIntensityMatrix`. Otherwise, returns a list of
\code{MALDIquant::MassPeaks} objects (default).}

\item{verbose:}{whether to print progress.}
}
\value{
A list of `MassPeaks` objects contaminated with noise.

A list of `MassPeaks` objects contaminated with noise, if \code{returnMat = FALSE)} or a sparse matrix of type \code{Matrix::dgCMatrix}
otherwise.
}
\description{
This contaminates MSI data with additive noise at certain m/z locations. This is used
for testing purposes.

This contaminates MSI data with additive noise at certain m/z locations. This is used
for testing purposes.
}
