% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task1.R
\name{task1_load_and_filter_data}
\alias{task1_load_and_filter_data}
\title{task 1: loading the data using DropletUtils, retaining only the protein coding genes in #' the data}
\usage{
task1_load_and_filter_data(gtf_path, matrix_zip_path, matrix_extract_dir)
}
\arguments{
\item{gtf_path}{the path to get to the gtf file}

\item{matrix_dir}{to extract the 10X matrix}
}
\value{
a filtered matrix only with the protein coding genes
}
\description{
This function extracts the data provvided and load them in the enviroment
it has a function to extract from gtf the geneid, gene biotype
to search for the protein coding genes. Filtering the matrix
}
\examples{
\dontrun{
gtf_path = "C:/Users/Utente/Desktop/magistrale/programming/exam/Homo_sapiens.GRCh38.111.g#' tf.gz",
matrix_zip_path = "C:/Users/Utente/Desktop/magistrale/programming/exam/filtered_feature_b#' c_matrix.zip",
matrix_extract_dir = "C:/Users/Utente/Desktop/magistrale/programming/exam/filtered_feat #' ure_bc_matrix/filtered_feature_bc_matrix"
)
sce <- result$sce
filtered_matrix <- result$filtered_matrix
dim(sce)
dim(filtered_matrix)
}
}
