% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task2.R
\name{task2_plot_genes_over_3_UMIs}
\alias{task2_plot_genes_over_3_UMIs}
\title{Title: distribution of genes with ≥3 UMIs}
\usage{
task2_plot_genes_over_3_UMIs(sce_object)
}
\arguments{
\item{sce_object}{loadeding SingleCellExperiment}
}
\value{
violin plot describing the distribution
}
\description{
here there is a function tant finds the genes with a ≥3 UMIs and displays it in a violin #' plot
}
\examples{
\dontrun{
task2_plot_genes_over_3_UMIs(filtered_matrix)
}
my_function(arg1, arg2)
}
