% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task3.R
\name{task3_filter_unwanted_genes}
\alias{task3_filter_unwanted_genes}
\title{filtering the gtf genes for ribosomial protein, ribosomial pseudogenes and mitoconrial
genes}
\usage{
task3_filter_unwanted_genes(counts_matrix, gtf)
}
\arguments{
\item{counts_matrix}{to have the counts of the genes}

\item{gtf}{it-d the data frame}
}
\value{
a summary table of the eliminated genes
}
\description{
searching for the ribosomial protein, ribosomial pseudogenes and mitoconrial
genes biotype, removing them
}
\examples{
\dontrun{
filtered_matrix_step3 <- result$filtered_matrix
print(result$summary_table)
}
}
