% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task4.R
\name{task4_run_pca_and_plot}
\alias{task4_run_pca_and_plot}
\title{running a PCA}
\usage{
task4_run_pca_and_plot(sce_object)
}
\arguments{
\item{sce_object}{to work with the 10X data}
}
\value{
the plot of the PCA
}
\description{
This function runs a PCA and plots it
}
\examples{
\dontrun{
result <- task4_run_pca_and_plot(filtered_matrix_step3)
sce <- result$sce
variance_explained <- result$variance_explained
}
}
