% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task5.R
\name{task5_run_umap_and_plot}
\alias{task5_run_umap_and_plot}
\title{running a UMAP based on the PCA results}
\usage{
task5_run_umap_and_plot(sce_object, n_pcs = 6)
}
\arguments{
\item{sce_object}{to work with the 10X data}

\item{n_pcs}{= 6}
}
\value{
the plot of the UMAP
}
\description{
This function runs a UMAP and plots it
}
\examples{
\dontrun{
result <- task5_run_umap_and_plot(sce, n_pcs = 6)
sce <- result$sce
umap_coords <- result$umap_coords
}
}
