% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task6.R
\name{task6_cluster_cells_and_plot}
\alias{task6_cluster_cells_and_plot}
\title{running a UMAP based on clusters}
\usage{
task6_cluster_cells_and_plot(sce_object, k = 10)
}
\arguments{
\item{sce_object}{to work with the 10X data}

\item{k=10}{parameters for the clusters}
}
\value{
the plot of the UMAP
}
\description{
This function runs a UMAP on clusters and plots them
}
\examples{
\dontrun{
result <- task6_cluster_cells_and_plot(sce, k = 10)
sce <- result$sce
clusters <- result$clusters
}
}
