% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task7.R
\name{task7_annotate_cells_with_singleR}
\alias{task7_annotate_cells_with_singleR}
\title{running a UMAP based on predicted clusters}
\usage{
task7_annotate_cells_with_singleR(sce_object)
}
\arguments{
\item{sce_object}{to work with the 10X data}
}
\value{
the plot of the UMAP
}
\description{
This function runs a UMAP on predicted clusters and plots them
}
\examples{
\dontrun{
result <- task7_annotate_cells_with_singleR(sce)
sce <- result$sce
predictions <- result$predictions
}
}
