% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task8.R
\name{task8_infer_tissue_origin}
\alias{task8_infer_tissue_origin}
\title{confronting the gene symbols of the gtf with a real tissue data base}
\usage{
task8_infer_tissue_origin(gtf_data, panglao_db)
}
\arguments{
\item{gtf_data}{my data}

\item{panglao_db}{external data}
}
\value{
the plots of the tissues
}
\description{
This function connects the gene symbols with the ones in the Panglao database
plotting all the tissues that found a match between the data sets and a plot
that shows only the top 25 tisuues most found
}
\examples{
\dontrun{
result <- task8_infer_tissue_origin(gtf, panglao)
}
}
