% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformIntensity.R
\name{transformIntensity}
\alias{transformIntensity}
\title{Intensity transformation for SPPs}
\usage{
transformIntensity(
  spp,
  method = "z-score",
  irange = c(0, 1),
  forceRange = TRUE
)
}
\arguments{
\item{spp:}{a spatial point pattern of type 'ppp' with its `marks` being a dataframe with
two columns `idx`and `intensity`.}

\item{method:}{a character specifying the method of transformation, one of
`c("z-score", "sqrt", "scaling")` such that `z-score` represents (analyte-wise)
standardization (subtraction of mean and division by std), `sqrt` is the square
root transformation and `scaling` is linear mapping of intensities into [0,1] range.}

\item{irange:}{a numeric vector of two elements speciying the output range of
the scalled intensities.}

\item{forceRange:}{logical, whether to force-map the intensities in to the range
given by `irange` regardless of the `method` used.}
}
\description{
Intensity transformation for SPPs
}
