% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwRapper.R
\name{unwRapper}
\alias{unwRapper}
\title{Unwrapper for feather files}
\usage{
unwRapper()
}
\value{
folder with optimized structures as xyz, and information needed for moleculaR.
}
\description{
Translate stored information into human readable csv files.
User is guided with questions.
}
