% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwRapper.R
\name{unwRapper.single}
\alias{unwRapper.single}
\title{Single molecule unwrapper for feather files}
\usage{
unwRapper.single(feather_file)
}
\arguments{
\item{feather_file}{A feather file resulting from extRactoR}
}
\value{
folder with optimized structures as xyz, and information needed for moleculaR.
}
\description{
Translate stored information into human readable csv files
}
\keyword{internal}
