% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cube_steRimol.R
\name{xyz.from.cube}
\alias{xyz.from.cube}
\title{Extract xyz file from Gausssian cube file}
\usage{
xyz.from.cube(cube_file)
}
\arguments{
\item{cube_file}{path to cube file}
}
\value{
write an xyz file with the same name
}
\description{
This function acts on Gaussian cube file, generating an xyz file
of the optimized structure
Input is the path to a .cube file
}
