% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{analytePointPattern}
\alias{analytePointPattern}
\title{analytePointPattern Class Constructor}
\usage{
analytePointPattern(
  spp = NA,
  x = NA,
  y = NA,
  win = NA,
  intensity = NA,
  mzVals,
  metaData = list()
)
}
\arguments{
\item{spp}{a spatial point pattern of type 'ppp' with its `marks` being a dataframe with
two columns `idx`and `intensity`.}

\item{x, y, intensity, win}{optional, either these arguments or 'spp' have to be supplied.}

\item{metaData}{a named list having additional attributes describing the individual masses in `mzVals`.}
}
\value{
An S3 object of type 'ppp' and 'analytePointPattern' with the following entries \cr
\itemize{
  \item regular `ppp` entries: see `spatstat` documentation for `ppp` class.
  \item metaData: A data frame with meta data describing the points in `spp$marks`.
}
}
\description{
This is a class to extend 'ppp' class only enforcing specific point-marks. It also
adds a 'metaData' slot to declare which analytes are constructing a given 'analytePointPattern' object.
It represents the spatial point pattern of an analyte (or set of
analytes) with marks being a data.frame of two columns c('idx', 'intensity'). This constructor is for
internal use and normally should not be called by the user.
}
\details{
`mzVals` and `metaData` are combined into a data.frame which becomes a class slot `metaData`.
This contains the `mzVals` and possible other meta data with each row having a unique identifier `idx`
which then link `metaData` with `spp$marks`.
}
\keyword{internal}
