% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_xyz_steRimol.R
\name{b1.scanner}
\alias{b1.scanner}
\title{scans for B1 vectors}
\usage{
b1.scanner(i, substi, plot.B1 = F)
}
\arguments{
\item{i}{degree}

\item{plot.B1}{should a plot of B1 and B5 box be presented}

\item{plane}{a 2D projection of the substituent's xyz, perpendicular to L axis}
}
\value{
list of B1 vectors
}
\description{
scans for B1 vectors
}
\keyword{internal}
