% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSparseMat.R
\name{createSparseMat}
\alias{createSparseMat}
\title{Creates a sparse matrix from a list of \code{MassPeaks} objects}
\usage{
createSparseMat(x)
}
\arguments{
\item{x:}{a list of \code{MassPeaks objects}.}
}
\value{
A sparse matrix, see \code{?Matrix::sparseMatrix} for more info.
}
\description{
The sparse matrix is created via \code{Matrix::sparseMatrix}.
}
\keyword{internal}
