% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{dip.gaussian.multi}
\alias{dip.gaussian.multi}
\title{Pulls and manipulates dipole moment vector.
Allows for use of several substructures.}
\usage{
dip.gaussian.multi(
  coor_atoms_vector = c(""),
  center_of_substructure = F,
  subunits_inputs_vector = NULL,
  center_of_mass = F
)
}
\arguments{
\item{center_of_substructure}{logical, should use center of mass of
a substructure as origin or not, if TRUE - center_of_mass must be FALSE}

\item{subunits_inputs_vector}{ONLY if center_of_substructure is TRUE - vector of
atoms character (e.g. c('10 11 12 13', '4 5 6 7 8 9'))}

\item{center_of_mass}{logical, should use center of mass of the
basic structure as origin or not, if TRUE - center_of_substructure
must be FALSE}

\item{coor_atoms}{3 or 4 atoms character (e.g. '1 2 3')}
}
\value{
data frame with multiple or single DMs
}
\description{
Pulls and manipulates dipole moment vector.
Allows for use of several substructures.
}
