% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dipoles.R
\name{dip.gaussian}
\alias{dip.gaussian}
\title{Use Gaussain's dipole moment, and perform manipulations to it.}
\usage{
dip.gaussian(
  coor_atoms = "",
  center_of_mass = F,
  center_of_mass_substructure = F,
  sub_atoms = NULL
)
}
\arguments{
\item{coor_atoms}{3 atoms character (e.g. '1 2 3')}

\item{center_of_mass}{logical, should use center of mass of the
basic structure as origin or not, if TRUE - center_of_mass_substructure
must be FALSE}

\item{center_of_mass_substructure}{logical, should use center of mass of
a substructure as origin or not, if TRUE - center_of_mass must be FALSE}

\item{sub_atoms}{ONLY if center_of_mass_substructure is TRUE - atoms character (e.g. '4 5 6 7 8 9')}
}
\value{
dipole moment
}
\description{
Use Gaussain's dipole moment, and perform manipulations to it.
}
\keyword{internal}
