% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{.bw.iterative}
\alias{.bw.iterative}
\title{Calculate Gaussian Bandwidth - `iterative`}
\usage{
.bw.iterative(
  spp,
  weighted = TRUE,
  bw = seq(0.5, 8, 0.5),
  csr = NULL,
  plot = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{spp:}{The spatial point pattern.}

\item{weighted:}{Whether the intensities are inlcude into the computation. Switch off for Collective Projections.}

\item{bw:}{A vector, The gaussian band width pool used for Kernel density estimation.}

\item{csr:}{Pre-computed wighted csr model corresponding to the given spp. This
speeds up the computation}

\item{plot:}{Whether to plot.}

\item{verbose:}{Whether to show progress.}
}
\value{
A list ..
}
\description{
This function calculates the Gaussian bandwidth to be used for analyte probability maps by
finding the curve infliction point for a curve that represents hotspot area as a function
of bandwidth. This is used internally in `moleculaR::probMap`.
}
\keyword{internal}
