% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateFwhm.R
\name{estimateFwhm}
\alias{estimateFwhm}
\title{Estimation of fwhm(m/z)}
\usage{
estimateFwhm(
  s,
  sampling = 1000L,
  plot = FALSE,
  savePlot = NULL,
  returnValues = FALSE
)
}
\arguments{
\item{s:}{a spectrum of \code{MALDIquant::Masss} type.}

\item{sampling:}{an integer specifying how many detected p to consider for fwhm estimation.}

\item{plot:}{whether to plot the result}

\item{savePlot:}{either \code{NULL} or file path to save a plot as svg.}

\item{returnValues:}{if set to \code{TRUE}, returns additionally the calculated fwhm values.}
}
\value{
Returns a linear interplotor function fwhm(m/z). If \code{returnValues = TRUE} a list is returned
storing the fwhm(m/z) interpolating function in addition to the calculated fwhm values.
}
\description{
This creates a linear interpolator function that approximats fwhm as a function
of m/z for the given single s.
}
