% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalspatialExpr.R
\name{evalSpatialExpr}
\alias{evalSpatialExpr}
\title{Evaluates an expression of SPP objects}
\usage{
evalSpatialExpr(exprn, dataList, ppwin, sqrtTransform = FALSE)
}
\arguments{
\item{expr:}{An unquoted (unwrapped) expression of 'ppp' objects listed in 'dataList'.}

\item{dataList:}{A named list containing the 'ppp' objects involved in 'expr'.}

\item{ppwin:}{The window 'spatstat::owin' object which all elements of 'dataList' share.}

\item{sqrtTransform:}{Optional square root transofmation of the 'ppp' objects in 'dataList'.}
}
\value{
A list containing 'spatstat::im' and the corresponding 'spatstat::ppp' objects resulting of
the applied expression 'expr'.
}
\description{
This evaluates an arbitrary expression involving spatial point pattern objects 'spatstat::ppp' which encode
analyte distributions of MSI data.
}
