% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractoR.R
\name{extractoR}
\alias{extractoR}
\alias{extRactoR.auto}
\title{Extract and compress all needed information from Gaussian log files}
\usage{
extractoR()
}
\value{
A .feather file for each log file
}
\description{
This function acts on a set of Gaussian log files.
No input is needed.
}
\details{
No Parameters
}
