% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy-initLipidSearch.R
\name{initLipidSearch}
\alias{initLipidSearch}
\title{Initialize lipids search}
\usage{
initLipidSearch(swissdb)
}
\arguments{
\item{swissdb:}{The swiss lipids db as dataframe.}
}
\value{
an S3 object with three entreis; a list containing lipid entries
organized into lipid classes, an additional empty 'hitslist'
with the same structure and naming to store the search results and
a character verctor summarizing all lipid classes in the first entry.
}
\description{
This function initializes some lists for lipid search.
}
