% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parallel_linear_models.R
\name{model.subset.parallel}
\alias{model.subset.parallel}
\title{Parallelized Subset Feature Selection with Correlation Threshold}
\usage{
model.subset.parallel(
  data,
  out.col = dim(data)[2],
  min = 2,
  max = floor(dim(data)[1]/5),
  results_name = "results_df",
  folds = nrow(data),
  iterations = 1,
  cutoff = 0.85,
  cor.threshold = 0.7
)
}
\arguments{
\item{data}{A data frame containing the features and the target column.}

\item{out.col}{An integer indicating the column index of the target variable (default: last column).}

\item{min}{The minimum number of features to include in a subset (default: 2).}

\item{max}{The maximum number of features to include in a subset (default: floor(nrow(data) / 5)).}

\item{results_name}{Name of file containing results. The function writes results on the fly, to free up RAM.}

\item{folds}{Number of folds for cross-validation (default: number of rows in `data`).}

\item{iterations}{Number of iterations for cross-validation (default: 1).}

\item{cutoff}{R-squared cutoff for model selection (default: 0.85).}

\item{cor.threshold}{Pearson correlation threshold for feature selection (default: 0.7).}
}
\value{
A data frame with the best models, their R-squared, Q-squared, and MAE values.
}
\description{
This function generates all possible subsets of features in a user-defined range (min to max),
ensuring that no two variables in a subset have a Pearson correlation coefficient above the
specified correlation threshold (`cor.threshold`). It performs parallel computation for faster execution.
}
