% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_models.R
\name{model.subset}
\alias{model.subset}
\title{Create Subsets of Features Based on Correlation Threshold and R-squared}
\usage{
model.subset(
  data,
  leave.out = "",
  out.col = dim(data)[2],
  min = 2,
  max = floor(dim(data)[1]/5),
  folds = nrow(data),
  iterations = 1,
  cutoff = 0.85,
  cor.threshold = 1
)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{leave.out}{A vector of row names to exclude from the data (default is an empty string).}

\item{out.col}{The index of the output column in the data (default is the last column).}

\item{min}{Minimum number of variables in a subset (default is 2).}

\item{max}{Maximum number of variables in a subset (default is floor(n/5)).}

\item{folds}{Number of folds for cross-validation (default is number of rows in data).}

\item{iterations}{Number of iterations for cross-validation (default is 1).}

\item{cutoff}{R-squared cutoff for model selection (default is 0.85).}

\item{cor.threshold}{Pearson correlation threshold to filter highly correlated variables (default is 0.7).}
}
\value{
A data frame containing the top 10 models based on R-squared, Q-squared, and MAE.
}
\description{
This function generates all possible subsets of features in a user-defined range (min to max), while ensuring
no two variables in a subset have a Pearson correlation coefficient of `cor.threshold` or higher. 
It returns the top models based on R-squared and further evaluates them using cross-validation to calculate Q-squared and MAE.
}
