% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moleculaR_main.R
\name{moleculaR.Input.Maker}
\alias{moleculaR.Input.Maker}
\title{User interface for the extraction of all possible features}
\usage{
moleculaR.Input.Maker()
}
\value{
csv file with all features and a input file for future
 use and documentation
}
\description{
Generate user inputs for future use.
The input can be used directly with the moleculaR function.
No input is needed. User is guided with questions and answers.
See the full manual for rules of use and options.
}
