% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moleculaR_main.R
\name{moleculaR}
\alias{moleculaR_main}
\title{User interface for the extraction of all possible features}
\usage{
moleculaR()
}
\value{
csv file with all features and a input file for future
 use and documentation
}
\description{
Calculate all features by answering questions, saving the final
data set in a chosen location and retaining user inputs for future use.
The output of inputs can be used directly with the moleculaR.input function.
No input is needed. User is guided with questions and answers.
See the full manual for rules of use and options.
}
