% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support_Functions.R
\name{name_changer}
\alias{name_changer}
\title{Change names of files in a directory, based on a matched pattern.}
\usage{
name_changer(dir, from, to)
}
\arguments{
\item{dir}{directory containing files}

\item{from}{pattern to change}

\item{to}{change pattern to}
}
\value{
changes names
}
\description{
Convenient wrapper for stringr::str_replace
}
