% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{plot.molProbMap}
\alias{plot.molProbMap}
\title{Plot molecular Probability Maps}
\usage{
\method{plot}{molProbMap}(
  obj,
  what = "detailed",
  densityAsBase = FALSE,
  sppArgs = list(),
  imageArgs = list(),
  fArgs = list(),
  rescale = TRUE,
  smooth = FALSE,
  zero.rm = FALSE,
  signifArea = "both",
  ionImage = NA,
  figLegend = TRUE,
  lwd.signifArea = 5,
  col.hotspot = "red",
  col.coldspot = "blue"
)
}
\arguments{
\item{obj}{S3 object of type `molProbMap`.}

\item{what}{What to plot, c("detailed", "analytePointPattern", "csrPointPattern", "analyteDensityImage",
"csrDensityImage", "kdeIntensitiesDistr", "MPM"). The "detailed" option plots all options.}

\item{densityAsBase}{a logical, if set to `TRUE`, the base image, ontop of which
hotspot/coldspot contours are plotted, is taken from the estimated density image
of m/z MOI. Otherwise (default), the Gaussian-weighted ion image is used.}

\item{sppArgs}{a named list of arguments to be passed to `plotAnalyte` or plotting
"analytePointPattern" and "csrPointPattern". There are sensible defaults.}

\item{imageArgs}{a named list of arguments to be passed to `spatstat.geom::plot.im` for plotting
"analyteDensityImage", "csrDensityImage" and "MPM". There are sensible defaults.}

\item{fArgs}{a named list of arguments to be passed to `base::plot` for plotting
"kdeIntensitiesDistr". There are sensible defaults.}

\item{rescale}{logical, whether to scale the intensities of the output plot
to [0,1] interval.}

\item{smooth}{logical, whether to apply image smoothing via an isotropic
Gaussian kernel.}

\item{zero.rm}{logical, whether to remove zero-valued pixels from the intensity
image.}

\item{signifArea}{a character indicating which significance area to plot i.e. c("both", "hotspot", "coldspot").}

\item{ionImage}{an optional rastered image of type `im` of the corresponding "regular"
ion image, used for comparison. Could be generated via `moleculaR::searchAnalyte(..., wMethod = "sum")`
and subsequently using `pixellate`.}

\item{figLegend}{logical, whether to show the hotspot and coldspot legends.}

\item{lwd.signifArea}{numeric, the width of the hotspot and coldspot contours.}

\item{col.hotspot}{a character specifying the color of the hotspot contour.}

\item{col.coldspot}{a character specifying the color of the coldspot contour.}
}
\value{
nothing, plots only.
}
\description{
A method to plot MPMs with different combinations.
}
