% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3classes.R
\name{plotAnalyte}
\alias{plotAnalyte}
\title{Plot analytePointPattern}
\usage{
plotAnalyte(
  obj,
  colourmap = "inferno",
  transpFactor = 0.7,
  pch = 19,
  size = 0.4,
  analyte = "m/z Analyte"
)
}
\arguments{
\item{obj}{S3 object of type `anlaytePointPattern` (and `spatstat.geom::spp`).}

\item{colourmap}{the colourmap to be used, see `?viridis::viridis_pal`.}

\item{transpFactor}{Transparency fraction. Numerical value or vector of values between 0 and 1,
giving the opaqueness of a colour. A fully opaque colour has `transpFactor=1`.}

\item{pch}{a positive integer, or a single character. See `?par`.}

\item{size}{the size of the symbol: a positive number or zero. see`?symbolmap`.}

\item{analyte}{character, name of the analyte.}
}
\value{
nothing, plots only.
}
\description{
A method to plot `anlaytePointPattern` objects. This calls `spatstat.geom::plot.ppp` with
pre-defined defaults. For more control over the plotting please use `spatstat.geom::plot.ppp`.
}
