% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{probMap}
\alias{probMap}
\title{Calculate molecular probablity maps}
\usage{
probMap(
  sppMoi,
  weighted = TRUE,
  control = NULL,
  bw = seq(1, 10, 1),
  bwMethod = "spAutoCor",
  csrIntensities = "resample",
  sqrtTansform = FALSE,
  csrMoi = NULL,
  pvalThreshold = 0.05,
  pvalCorrection = "bonferroni",
  bwPlot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{sppMoi:}{The spatial point pattern.}

\item{weighted:}{Whether the intensities are inlcude into the computation.}

\item{bw:}{The Gaussian bandwidth used for Kernel density estimation. If a single numeric value is supplied
then it will be used for the computation of KDE. If a vector is supplied (in increasing order), then these values
will be passed to internal methods `spAutoCor` or `iterative` for computing the bandwidth based on cross validation,
see `?moleculaR::.bw.spAutoCorr` and `?moleculaR::.bw.iterative`.}

\item{bwMethod:}{The method used for computing the Gaussian bandwidth, c("spAutoCor","iterative", "scott").}

\item{bwPlot:}{whether to plot the Gaussian bw selection procedure, ignored when \code{bwMethod = "scott"}.}

\item{csrIntensities:}{How the intensities for the csrMoi model are generated, c("resample", "Poisson", "Gaussian").}

\item{sqrtTansform:}{whether to apply square root transformation to the intensities of `sppMoi`. Set to 'TRUE' when
'sppMoi' cotains a set of different analytes as in the case of collective projection maps.}

\item{control:}{An sppMoi object that is designated as the "control" or "null" alternative of \code{sppMoi}. If supplied the
csrMoi model will be generated from the intensities of this object.}

\item{csrMoi:}{Pre-computed wighted csrMoi model corresponding to the given sppMoi. This
could be useful when generating hotspot-bw curves.}

\item{pvalThreshold:}{The p-value threshold to be used for the hypothesis testing.}

\item{pvalCorrection:}{The method used for p-values correction, see '?p.adjust' for more details.}

\item{verbose:}{whether to output progress.}
}
\value{
#' A list of \cr
\itemize{
  \item GausBW: calculated Gaussian bandwidth.
  \item sppMoi: the input spp with intensities sqrt-transformed if 'sqrtTransform == TRUE'.
  \item csrMoi: calculated complete spatial randomness model for the input sppMoi.
  \item rhoMoi: density image of the input point pattern sppMoi.
  \item rhoCsr: density image of the computed csrMoi pattern.
  \item hotspotpp: the remaining points (type 'ppp') which lie within the hotspot mask as a spatial point pattern.
  \item hotspotIm: the hotspot image of type 'im'.
  \item hotspotMask: the hotspot mask of type 'owin'.
  \item coldspotpp: the remaining points (type 'ppp') which lie within the coldspot mask as a spatial point pattern.
  \item coldspotIm: the coldspot image of type 'im'.
  \item coldspotMask: the hotspot mask of type 'owin'.
}
}
\description{
This function creates molecular probability maps from a given spatial point pattern representation.
}
