% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{probMap}
\alias{probMap}
\title{Calculate molecular probablity maps}
\usage{
probMap(
  spp,
  win,
  weighted = TRUE,
  bw = NULL,
  control = NULL,
  bwMethod = "iterative",
  csr = NULL,
  bwPlot = FALSE
)
}
\arguments{
\item{spp:}{The spatial point pattern.}

\item{win:}{The window object of type `spatstat::owin`.}

\item{weighted:}{Whether the intensities are inlcude into the computation.}

\item{bw:}{The Gaussian bandwidth used for Kernel density estimation.}

\item{bwMethod:}{The method used for computing the Gaussian bandwidth, c("iterative", "scott").}

\item{bwPlot:}{whether to plot the Gaussian bw selection procedure, ignored when \code{bwMethod = "scott"}.}

\item{control:}{An spp object that is designated as the "control" or "null" alternative of \code{spp}. If supplied the
CSR model will be generated from the intensities of this object.}

\item{csr:}{Pre-computed wighted csr model corresponding to the given spp. This
could be useful when generating generating hotspot-bw curves.}
}
\value{
A list of
cutoff: the chosen threshold above which it is a hotspot,
GausBW: calculated Gaussian bandwidth,
csrspp: calculated complete spatial randomness model for the input spp,
denspp: density image of the input point pattern spp,
denCsr: density image of the computed csr pattern,
hotspotpp: the remaining points which lie within the hotspot mask as a spatial point pattern,
hotspotIm: the hotspot image,
hotspotMask: the hotspot mask of type owin.mask,
nonHotspotMask: the non-hotspot areas, to be used to mask the non hotspot areas.
}
\description{
This function creates molecular probability maps from a given spatial point pattern representation.
}
