% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMap.R
\name{probMap}
\alias{probMap}
\title{Calculate molecular probablity maps}
\usage{
probMap(
  sppMoi,
  reference = NULL,
  bw = seq(0.5, 10),
  edgeCorrection = FALSE,
  bwMethod = "spAutoCor",
  csrMoi = NULL,
  pvalThreshold = 0.05,
  pvalCorrection = "BH",
  diagPlots = FALSE,
  seed = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{sppMoi}{The spatial point pattern.}

\item{reference}{An sppMoi object that is designated as the "control" or "null" alternative of \code{sppMoi}. If supplied the
csrMoi model will be generated from the intensities of this object.}

\item{bw}{The Gaussian bandwidth used for Kernel density estimation. If a single numeric value is supplied
then it will be used for the computation of KDE. If a vector is supplied (in increasing order), then these values
will be passed to internal method `spAutoCor` for computing the bandwidth based on cross validation,
see `?moleculaR::.bw.spAutoCorr`.}

\item{edgeCorrection}{a logical. Whether to apply edge correction through
erosion by calling `spatstat.geom::erosion`. This could be beneficial for
removing edge artifacts which might otherwise impact the result.}

\item{bwMethod}{The method used for computing the Gaussian bandwidth, c("spAutoCor", "scott").}

\item{csrMoi}{Pre-computed wighted csrMoi model corresponding to the given sppMoi. This
could be useful when generating hotspot-bw curves. For internal use only.}

\item{pvalThreshold}{The p-value threshold to be used for the hypothesis testing.}

\item{pvalCorrection}{The method used for p-values correction, see '?p.adjust' for more details.}

\item{diagPlots}{whether to plot the Gaussian bw selection procedure, ignored when \code{bwMethod = "scott"}.}

\item{seed}{a single value, interpreted as integer or `NULL` (default) controlling the process of random
number generation. If set, ensures that the same CSR model is generated at different runs for the
same input `sppMoi`}

\item{verbose}{whether to output progress.}

\item{csrIntensities}{How the intensities for the csrMoi model are generated, c("resample", "Poisson", "Gaussian").}

\item{...:}{arguments passed to `plot` for bandwidth plotting when `bwMethod="spAutoCor"`.}
}
\value{
#' An S3 object of type molProbMap with the following entries \cr
\itemize{
  \item bw: calculated Gaussian bandwidth.
  \item sppMoi: the input spp with intensities sqrt-transformed if 'sqrtTransform == TRUE'.
  \item csrMoi: calculated complete spatial randomness model for the input sppMoi.
  \item rhoMoi: density image of the input point pattern sppMoi.
  \item rhoCsr: density image of the computed csrMoi pattern.
  \item hotspotpp: the remaining points (type 'ppp') which lie within the hotspot mask as a spatial point pattern.
  \item hotspotIm: the hotspot image of type 'im'.
  \item hotspotMask: the hotspot mask of type 'owin'.
  \item coldspotpp: the remaining points (type 'ppp') which lie within the coldspot mask as a spatial point pattern.
  \item coldspotIm: the coldspot image of type 'im'.
  \item coldspotMask: the hotspot mask of type 'owin'.
}
}
\description{
This function creates molecular probability maps from a given spatial point pattern representation.
}
