% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCentrData.R
\name{readCentrData}
\alias{readCentrData}
\title{Read imzML data}
\usage{
readCentrData(path, verbose = FALSE, ...)
}
\arguments{
\item{path:}{character string, path to the imzML file which should be read in.}

\item{verbose:}{logical, verbose output?}

\item{...:}{other arguments passed to \code{MALDIquantForeign::importImzMl}.}
}
\value{
A list of \code{MALDIquant::MassPeaks} objects if the input file is of 'processed' type.
}
\description{
A wrapper for \code{MALDIquantForeign::importImzMl} for reading centroided (i.e. processed)
and continuous imzML data. Note that to use this function, \code{MALDIquantForeign} has 
to be installed separately, i.e. \code{MALDIquantForeign} is not installed automatically with 
\code{moleculaR}.
}
