% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{ring.vib.df}
\alias{ring.vib.df}
\title{Get a ring's characteristic vibrations frequencies, for a set of molecules}
\usage{
ring.vib.df(input = NA)
}
\arguments{
\item{input}{The atom index of the first ring atom bonded to the basic
strcuture.}
}
\value{
A data frame with stretching frequencies
}
\description{
Get a ring's characteristic vibrations frequencies, for a set of molecules
}
