% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{ring.vib.multi}
\alias{ring.vib.multi}
\title{Get a ring's characteristic vibrations frequencies}
\usage{
ring.vib.multi(inputs_vector)
}
\arguments{
\item{inputs_vector}{vector of atom indices of the first ring atom bonded to the basic
strcuture, for each ring.}
}
\value{
A data frame with stretching frequencies
}
\description{
used on a set of molecules, applicable for as many rings wanted
}
