% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vibrations.R
\name{ring.vib}
\alias{ring.vib}
\title{Get a ring's characteristic vibrations frequencies}
\usage{
ring.vib(primary_atom)
}
\arguments{
\item{primary_atom}{The atom index of the first ring atom bonded to the basic
strcuture.}
}
\value{
A data frame with stretching frequencies
}
\description{
Get a ring's characteristic vibrations frequencies
}
\keyword{internal}
