% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAnalyte.R
\name{searchAnalyte}
\alias{searchAnalyte}
\title{Find an analyte in the dataset}
\usage{
searchAnalyte(m, fwhm, massAxis, spData, coords, wMethod = "Gaussian")
}
\arguments{
\item{m:}{the mass to be queried (Da).}

\item{fwhm:}{the fwhm at `m`.}

\item{massAxis:}{the mass axis of the whole MS dataset to be searched against.}

\item{spData:}{the sparse matrix containing the data.}

\item{coords:}{the coordinates of spectra contained within the spData.}

\item{wMethod:}{wighting method; c("Gaussian", "sum", "max", "mean").}
}
\value{
A dataframe containing the hits for `m` in `msData`.
}
\description{
This function check if an analyte (given by m/z) is detectable in the MS data represented 
by a sparse matrix
}
