% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wVDW_steRimol.R, R/get_xyz_steRimol.R
\name{steRimol.xyz.multi}
\alias{steRimol.xyz.multi}
\title{Multi-Axis Sterimol Parameter Calculation - xyz}
\usage{
steRimol.xyz.multi(coordinates_vector, CPK = T, only_sub = T, drop = NULL)

steRimol.xyz.multi(coordinates_vector, CPK = T, only_sub = T, drop = NULL)
}
\arguments{
\item{coordinates_vector}{primary axes, a vector of two atom characters}

\item{CPK}{logical, if TRUE will use CPK radii based on Verloop's atom type,
if FALSE will use Pyykko's covalent radii}

\item{only_sub}{if TRUE (default) will account only for atoms directly bonded
from the primary axis onward}

\item{drop}{numeric value of an atom, from which onward atoms will be dropped from calculation}
}
\value{
Data frame containing Sterimol parameters for each molecule and axis combination

sterimol values (multiple)
}
\description{
Calculates charge weighted VDW radii Sterimol parameters 
for multiple primary axes across multiple molecules.

Calculate multiple sterimol values for a set of molecules
}
\details{
Enables calculation of Sterimol parameters for multiple substituents or substructures
in a set of molecules, using derived van der Waals radii.
}
