% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superImposeAnalytes.R
\name{superimposeAnalytes}
\alias{superimposeAnalytes}
\title{Superimpose analytePointPatterns}
\usage{
superimposeAnalytes(pppObjs, spWin = NULL, check = TRUE)
}
\arguments{
\item{pppObjs}{a list of any number of `ppp` and `analytePointPatterns` objects.}

\item{spWin}{the spatial window where all points will be projected.}

\item{check}{logical, passed to `spatstat.geom::superimpose.ppp`.}
}
\value{
An analyte point patter of type `ppp` and `analytePointPattern` containing all supplied analytes.
}
\description{
Superimpose any number of point patterns of type `ppp` and `analytePointPatterns`.
}
